% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akaike.weight.R
\name{akaike.weight}
\alias{akaike.weight}
\title{Compute akaike weights}
\usage{
akaike.weight(x, measure = c("aic", "loglikelihood"))
}
\arguments{
\item{x}{A numeric vector or matrix with the AIC or log likelihood values}

\item{measure}{A string (default \code{"aic"}) with the type of value in x, either \code{"aic"} or \code{"loglikelihood"}; can be abbreviated.}
}
\description{
Compute akaike weights
}
\examples{
AICvalues <- c(204,202,206,206,214)
akaike.weight(AICvalues)
}
\references{
Wagenmakers, E., & Farrell, S. (2004). AIC model selection using Akaike weights. Psychonomic Bulletin & Review, 11(1), 192–196. https://doi.org/10.3758/BF03206482
}
