% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akaike_weight.R
\name{akaike_weight}
\alias{akaike_weight}
\title{Compute akaike weights}
\usage{
akaike_weight(x, measure = c("aic", "loglikelihood"))
}
\arguments{
\item{x}{A numeric vector or matrix with the AIC or log likelihood values}

\item{measure}{A string (default \code{"aic"}), the type of values in \code{x}. Can be one of \code{"aic"}, \code{"loglikelihood"}; can be abbreviated.}
}
\description{
Compute akaike weights
}
\examples{
AICvalues <- c(204,202,206,206,214)
akaike.weight(AICvalues)
}
