% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm-methods-anova.R
\name{anova.cm}
\alias{anova.cm}
\title{Analysis of Deviance for Cogscimodel Fits}
\usage{
\method{anova}{cm}(object, ...)
}
\arguments{
\item{object, ...}{Objects of class \code{cm}, typically the result of a call to one fo the models. (Or a list of \code{objects} for the \code{"cm_list"} method.)}
}
\value{
An anova-style table
}
\description{
\code{anova} returns an analysis of deviance table for one or more fitted cognitive models
}
\examples{
# None yet.
}
