% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R, R/anova.cogscimodel.R
\name{anova.cogscimodel}
\alias{anova.cogscimodel}
\title{Analysis of Deviance for Cogscimodel Fits}
\usage{
\method{anova}{cogscimodel}(object, ...)

\method{anova}{cogscimodel}(object, ...)
}
\arguments{
\item{object, ...}{Objects of class \code{cogscimodel}, typically the result of a call to one fo the models. (Or a list of \code{objects} for the \code{"cogscimodellist"} method.)}
}
\value{
An anova-style table

An anova-style table
}
\description{
`anova` returns an analysis of deviance table for one or more fitted cogscimodels.

`anova` returns an analysis of deviance table for one or more fitted cogscimodels.
}
\examples{
None yet.
None yet.
}
