% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chr_as_rhs}
\alias{chr_as_rhs}
\title{Convert a character to formula
Checks if x is a character and if so converts it into RHS formula}
\usage{
chr_as_rhs(x)
}
\description{
Convert a character to formula
Checks if x is a character and if so converts it into RHS formula
}
