% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm-class.R
\name{Cm}
\alias{Cm}
\alias{cm-class}
\title{The R6 class underlying all "cm" (cognitive model) objects}
\description{
The R6 class underlying all "cm" (cognitive model) objects

The R6 class underlying all "cm" (cognitive model) objects
}
\details{
\code{Cm$new(formula, data, parspace)}

\code{parspace}. It is optional to define a value that makes the parameter have zero effect in the column called "na" in \code{parspace}. For example, a parameter \code{b} in \code{b*x}, has no-effect  when setting \code{b=0}. The no-effect value can be \code{NA}.

\bold{fix}
Supplying \code{list(alpha=0.5, beta="alpha", gamma=NA)} constrains the parameters alpha, beta, and gamma as follows
\itemize{
\item{\code{alpha=0.5}}{: fix alpha to 0.5.}
\item{\code{beta="alpha"}}{: fix beta to the value of alpha; alpha may be a free or fixed model parameter in \code{parspace}.}
\item{\code{delta=NA}}{: ignore delta if it can be ignored by setting delta equal to the value in the column "na" of \code{parspace}, given that parspace has a value in the column "na" for delta.}
}
You can ignore a model parameter by setting it to \code{NA} in \code{fix}, in this case your \code{parspace} needs to contain a value in the column na nullifying the effect of the parameter.
}
\section{Options, see \link{cm_options}, possible options}{

\describe{
\item{\code{fit}}{(default \code{TRUE}), \code{FALSE} omits fitting the free parameter.}
\item{\code{fit_measure}}{(default \code{"loglikelihood"}). When fitting, which fit measure to use? See \link[cognitiveutils]{gof}'s argument \code{type}.}
\item{\code{fit_data}}{(default: \code{data}). When fitting, which data other than to fit the model to? Useful if you fit to other data than you predict.}
\item{\code{fit_n}}{(default: \code{1}). When fitting, the number of reservations underlying each data point in \code{fit_data}. If this data is aggregated data, change this argument.}
\item{\code{fit_roi_solver}}{(default: \code{"auto"}). When fitting with the \link{http://roi.r-forge.r-project.org/index.html}{ROI pakage}, which solver to use for the parameter estimation problem.}
\item{\code{nbest}}{(default: \code{self$npar("free")}, no. of free parameters) When fitting with grid search followed by a solver, how many best grid-search solutions to use as starting parameters with the solver?}
\item{\code{fit_options}}{(default: \code{NULL}). Other options, see \link[cognitiveutils]{gof}.}
}
}

\examples{
# No examples yet.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{call}}{The call to the model}

\item{\code{title}}{A string, the name of the model}

\item{\code{formula}}{Calling formula, e.g., \code{y ~ x1 + x2 + x3} models one response variable (\code{y}) given one three-attribute stimulus, and \code{y ~ x1 + x2 | z1 + z2} models a response given two two-attribute stimuli where attributes are separated by a "\code{+}"" and stimuli are separated by a "\code{|}".}

\item{\code{input}}{Inputs to the model in a standardized form: three-dimensional matrix based on \code{formula} and \code{data} with dimensions n(data) x n(attributes) x n(stimuli)}

\item{\code{more_input}}{Additional inputs to the model from variables that are not part of the formula, three-dimensional matrix with dimensions n(data) x n(additional inputs) x n(stimuli)}

\item{\code{res}}{The response variable to be modeled, two-dimensional matrix with dimensions n(data) x n(response variables)}

\item{\code{natt}}{List with the number of attributes per stimulus}

\item{\code{nobs}}{Number of observations in \code{data}}

\item{\code{nstim}}{Number of stimuli in model input}

\item{\code{nres}}{Number of response variables}

\item{\code{ncon}}{Number of constraints on the parameter}

\item{\code{par}}{Model parameters, a named list}

\item{\code{fix}}{Fixed model parameters, a named list}

\item{\code{parspace}}{Parameter space, a matrix with one parameter per row, and four columns lb (lower bound), ub (upper bound), start (starting value for fitting) and na (value of the parameter if it has no effect on the model); rownames are parameter names}

\item{\code{pred}}{Model predictions for \code{data}}

\item{\code{parnames}}{Parameter names}

\item{\code{stimnames}}{Stimuli names}

\item{\code{prednames}}{Predictio names}

\item{\code{mode}}{A string the modality of the predicted responses "\code{continuous}"" or "\code{discrete}"}

\item{\code{choicerule}}{A string with the choicerule that the model uses (e.g. "\code{softmax}"")}

\item{\code{pred_types}}{The types of predictios the model males}

\item{\code{discount}}{A numeric with trials to discount when fitting}

\item{\code{constraints}}{An object with the parameter constraints}

\item{\code{fitobj}}{An object holding the results of the parameter fitting method}

\item{\code{options}}{A list with options}

\item{\code{pass_checks}}{A logical if \code{TRUE} the model passes all internal checks}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Cm$new()}}
\item \href{#method-fit}{\code{Cm$fit()}}
\item \href{#method-coef}{\code{Cm$coef()}}
\item \href{#method-predict}{\code{Cm$predict()}}
\item \href{#method-simulate}{\code{Cm$simulate()}}
\item \href{#method-set_data}{\code{Cm$set_data()}}
\item \href{#method-set_par}{\code{Cm$set_par()}}
\item \href{#method-get_par}{\code{Cm$get_par()}}
\item \href{#method-npar}{\code{Cm$npar()}}
\item \href{#method-gof}{\code{Cm$gof()}}
\item \href{#method-logLik}{\code{Cm$logLik()}}
\item \href{#method-BIC}{\code{Cm$BIC()}}
\item \href{#method-AIC}{\code{Cm$AIC()}}
\item \href{#method-AICc}{\code{Cm$AICc()}}
\item \href{#method-MSE}{\code{Cm$MSE()}}
\item \href{#method-SSE}{\code{Cm$SSE()}}
\item \href{#method-RMSE}{\code{Cm$RMSE()}}
\item \href{#method-set_lb}{\code{Cm$set_lb()}}
\item \href{#method-set_ub}{\code{Cm$set_ub()}}
\item \href{#method-set_start}{\code{Cm$set_start()}}
\item \href{#method-print}{\code{Cm$print()}}
\item \href{#method-summary}{\code{Cm$summary()}}
\item \href{#method-pargrid}{\code{Cm$pargrid()}}
\item \href{#method-getCall}{\code{Cm$getCall()}}
\item \href{#method-clone}{\code{Cm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initializes a new model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cm$new(
  formula,
  data = NULL,
  parspace = make_parspace(),
  fix = NULL,
  choicerule = if (mode == "continuous") {     "none" } else {     NULL },
  title = NULL,
  discount = NULL,
  mode = NULL,
  options = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formula}}{A formula, e.g., \code{y ~ x1 + x2}.}

\item{\code{data}}{A data frame containing \code{formula}'s variables.}

\item{\code{parspace}}{(optional, \bold{required} to add model parameters) A n x 4 matrix, the parameter space. Use \link{make_parspace} to construct it. Column names must be \verb{"lb","ub","start","na"}, row names must be parameter names. Columns contain the lower limit, upper limit, starting value in fitting, and (optional) a value that makes a parameter have zero effect, which can be NA. See details.}

\item{\code{fix}}{(optional) Parameter constraints. Can be \code{"start"} or a list with \code{parname=value}-pairs. Parameter names see below under "parameter space".
\itemize{
\item{\code{"start"} constrains all available parameters to their starting values. Useful for model testing.}
\item{\code{parname = 0.5} constrains a parameter to 0.5.}
\item{\code{parname = "p2"} constrains a parameter to another model parameter \code{p2}.}
\item{\code{parname = NA} tells the model to omit a parameter, if possible.}
}}

\item{\code{choicerule}}{(only for discrete models) A string, the choice rule, to view, rum \code{cm_choicerules()}.}

\item{\code{title}}{(optional, default is the class name) A string, the model's name.}

\item{\code{discount}}{(optional) An integer or integer vector (default \code{0}), ddefining which or how many, starting from trial 1, to discount when fitting.}

\item{\code{mode}}{A string, the response mode. Allowed are \code{"discrete"}, `"continuous". Discrete responses are binary (0 or 1), continuous responses are numbers-}

\item{\code{options}}{(optional) Options to control the parameter fitting methods, see the "Options" section of \code{\link{cm_options}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\subsection{Method \code{fit()}}{
Fits the free model parameters to the response data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cm$fit(solver = self$options$solver, measure = self$options$fit_measure, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{solver}}{(optional) A string with the model solver}

\item{\code{measure}}{(optional) A string with the goodness-of-fit measure that the solver optimizes (e.g. \code{"loglikelihood"}). Possible values, see the \code{type} argument in \link[cognitiveutils]{gof}}

\item{\code{...}}{other arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-coef"></a>}}
\subsection{Method \code{coef()}}{
Returns the free parameters of the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cm$coef()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\subsection{Method \code{predict()}}{
Make predictions for every row in the input data to the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cm$predict(type = "response", newdata = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(optional, default \code{"response"}) Type of prediction to make. See the respective model.}

\item{\code{newdata}}{(optional) A data frame with new data for which to compute predictions.}

\item{\code{...}}{other arguments (ignored)
calls the child model's make_prediction function for
every slice of the third dimension of the input matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-simulate"></a>}}
\subsection{Method \code{simulate()}}{
Simulates observable data from the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cm$simulate(newdata = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newdata}}{(optional) A data frame with inputs (e.g., stimuli) to be used in the simulation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A matrix with simulated choices or judgment values
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_data"></a>}}
\subsection{Method \code{set_data()}}{
New data input for a cogscim
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cm$set_data(data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A data frame with variables corresponding to the inputs that the model needs}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_par"></a>}}
\subsection{Method \code{set_par()}}{
Set model parameter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cm$set_par(x, check = TRUE, constrain = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A named list of parameters, their names must be in \code{parspace}}

\item{\code{check}}{(optional, default \code{TRUE}) A logical value, if \code{TRUE} the parameter names are checked for validity, if \code{FALSE} no check is performed.}

\item{\code{constrain}}{(optional, default \code{TRUE}) A logical, whether to newly constrain the parameters when setting a parameter that is constrained.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_par"></a>}}
\subsection{Method \code{get_par()}}{
Get the model parameter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cm$get_par(x = "all")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(optional) A string, which set of parameters to get, \code{"all"} returns all parameter, \code{"free"} returns the free parameters, \code{"constrained"} returns constrained parameters, \code{"equal"} returns parameters equal to another parameter}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-npar"></a>}}
\subsection{Method \code{npar()}}{
Number of model parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cm$npar(x = "free")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A string, which of the parameters to return, allowed are \code{"all", "free", "constrained", "equal"}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gof"></a>}}
\subsection{Method \code{gof()}}{
Computes the goodness of model fit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cm$gof(
  type,
  n = self$options$fit_n,
  newdata = self$options$fit_data,
  discount = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{A string, which goodness of fit measure to use, e.g. \code{"loglikelihood"}; for the allowed values see \link[cognitiveutils]{gof}}

\item{\code{n}}{(optional) When fitting to aggregate data, supply how many raw data points underly each aggregated data point}

\item{\code{newdata}}{(optional) A data frame with new data - experimental!}

\item{\code{discount}}{(optional) An integer or integer vector (default \code{0}), ddefining which or how many, starting from trial 1, to discount when fitting.}

\item{\code{...}}{other arguments (ignored)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logLik"></a>}}
\subsection{Method \code{logLik()}}{
Log likelihood of the observed responses under the model predictions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cm$logLik(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{other arguments (ignored)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BIC"></a>}}
\subsection{Method \code{BIC()}}{
Bayesian Information Criterion of the model predictions given the observed responses
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cm$BIC(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{other arguments (ignored)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AIC"></a>}}
\subsection{Method \code{AIC()}}{
Akaike Information Criterion of the model predictions given the observed response
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cm$AIC(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{other arguments (ignored)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AICc"></a>}}
\subsection{Method \code{AICc()}}{
Small-sample corrected Akaike Information Criterion of the model predictions given the responses
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cm$AICc(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{other arguments (ignored)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MSE"></a>}}
\subsection{Method \code{MSE()}}{
Mean squared error between the model predictions given the observed responses
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cm$MSE(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{other arguments (ignored)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SSE"></a>}}
\subsection{Method \code{SSE()}}{
Sum of squared errors between the model predictions given the observed responses
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cm$SSE(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{other arguments (ignored)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RMSE"></a>}}
\subsection{Method \code{RMSE()}}{
Root mean squared error between the model predictions and the observed responses
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cm$RMSE(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{other arguments (ignored)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_lb"></a>}}
\subsection{Method \code{set_lb()}}{
Change the lower limit of parameter values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cm$set_lb(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{new parameter bound, e.g. \code{alpha = 0} sets the lower bound of the parameter \code{alpha} to 0}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_ub"></a>}}
\subsection{Method \code{set_ub()}}{
Change the upper limit of parameter values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cm$set_ub(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{new parameter bound, e.g. \code{alpha = 0} sets the lower bound of the parameter \code{alpha} to 0}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_start"></a>}}
\subsection{Method \code{set_start()}}{
Change the starting values for fitting parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cm$set_start(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{new parameter bound, e.g. \code{alpha = 0} sets the lower bound of the parameter \code{alpha} to 0}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Prints the model object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cm$print(digits = 2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{digits}}{A number specifying the number of digits to print}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\subsection{Method \code{summary()}}{
Summarizes the model, it's parameters, and some fit information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cm$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pargrid"></a>}}
\subsection{Method \code{pargrid()}}{
Produces a parameter grid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cm$pargrid(nsteps = NULL, offset = NULL, par = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nsteps}}{A list with steps for each parameter}

\item{\code{offset}}{A number by which to offset the parameters from their bounds}

\item{\code{par}}{(optional) A vector with parameter names for which to generate the grid.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCall"></a>}}
\subsection{Method \code{getCall()}}{
Retrieves the call
Remove choicerule
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cm$getCall()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
