% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cm_options.R
\name{cm_choicerules}
\alias{cm_choicerules}
\title{Show the Choicerules for Discrete Cognitive Models}
\usage{
cm_choicerules()
}
\description{
Show the Choicerules for Discrete Cognitive Models
}
\examples{
cm_choicerules()
}
