% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cm_options.R
\name{cm_options}
\alias{cm_options}
\title{Advanced Options for Cognitive Models}
\usage{
cm_options(
  lb = NULL,
  ub = NULL,
  start = NULL,
  fit = TRUE,
  fit_measure = "loglikelihood",
  fit_n = 1L,
  fit_data = NULL,
  solver = "auto",
  fit_grid_offset = 0L,
  fit_control = list(),
  fit_args = list()
)
}
\arguments{
\item{lb}{A named numeric vector, minimum values that a parameter may take. E.g., \code{c(k = -10)} lets a parameter \emph{k} start at -10.}

\item{ub}{A named numeric vector, maximum value that parameters may take. E.g., \code{c(k = 10)} lets a parameter \emph{k} go until 10.}

\item{fit}{Logical (default \code{TRUE}), \code{FALSE} disables parameter fitting. Useful for testing models.}

\item{fit_measure}{A string (default \code{"loglikelihood"}), the goodnes of fit measure that is optimized during parameter estimation. Can be one of the \code{types} in the function \link[cognitiveutils]{gof}:
\itemize{
\item \code{"loglikelihood"} is log likelihood which uses a
\itemize{
\item ... binomial PDF in discrete-data models.
\item ... normal PDF in continuous-data models, this adds a free parameter \code{"sigma"}, the standard deviation of the PDF, estimated from the data.
}
\item \code{"mse"} is mean-squared error
\item \code{"accuracy"} is percent accuracy
}}

\item{fit_n}{An integer, the number of data points underlying aggregated data. If the data that is being predicted is aggregated data, \code{n} is the number observations underlying each data point in the aggregated data.}

\item{fit_data}{A data frame, the data to estimate the model parameters from. Needed if the data for the parameter estimation differs from the data in the main \code{data} argument in a model.}

\item{solver}{A string, the alorithm used for parameter estimation,  \link{cm_solvers()} lists the options. Changing this may cause warnings about ignored options and may cause parameter bounds to be ignored and the model to fail.
\itemize{
\item \code{"grid"} uses a grid-search
\item \code{"solnp"} uses \link[Rsolnp]{solnp}
\item ... and 21 other solvers such as \code{"optimx"}, \code{"nloptr"}, and \code{"nlminb"} from \link{ROI} - R optimization infrastructure
\item \code{c("grid", "abcde")}, where abcde is a solver, performs a grid search, followed by an optimization with the solver. The solver optimizes n times, using the n best parameter sets that resulted from the grid search as start values. The overal best parameter result wins.
}}

\item{fit_grid_offset}{A small number, amount to offset the parameter values in a grid search from the parameter boundaries.}

\item{fit_control}{A list, arguments passed to the \strong{\code{control}} argument in the solver function, allowed values depend on the \code{solver} that you want to use to for the control options.}

\item{fit_args}{A list, additional arguments passed \strong{directly} to the solver function, see the respective pages of the solver that you want to use, to see which arguments the solver function has.}
}
\description{
Advanced Options for Cognitive Models
}
