% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cm_options.R
\name{cm_solvers}
\alias{cm_solvers}
\title{Show the Optimization Solvers for Cognitive Models}
\usage{
cm_solvers()
}
\description{
Show the Optimization Solvers for Cognitive Models
}
\examples{
cm_solvers()
}
