% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitivemodel-class.R
\name{cognitivemodel}
\alias{cognitivemodel}
\title{Add components to a model}
\usage{
cognitivemodel(data, ...)
}
\arguments{
\item{data}{a data.frame holding the variables that the first model uses}
}
\description{
\code{cogscimodel()} initializes a cogscimodel object. It can be used to
declare the input data being used by the model.
}
\details{
\code{cogscimodel()} is used to construct the initial model object,
and is almost always followed by \verb{\%+\%} to add a component to the
model.
}
