% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_init.R
\docType{package}
\name{cognitivemodels-package}
\alias{cognitivemodels}
\alias{cognitivemodels-package}
\title{cognitivemodels: Cognitive Models - Estimation, Prediction, and Development of Models for Cognitive Scientists}
\description{
Implementation of cognitive models, includes a general class cognitive models that has methods to fit free parameters to data and, can apply choicerules, and various goodness-of-fit measures (loglikelihood, MSE, etc.). The interface is similar to the lm() interface.
}
\details{
The following cogniitve models are part of this packge:
\itemize{
\item baseline - constant or 1/N choice models (as sanity check)
\item bayes - cognitive bayesian updating model
\item cpt - cummulative prospect theory (Tversky & Kahneman, 1992)
\item ebm - exemplar-based models
\item hm1988 - Houston & McNamara's optimal model for risk-sensitive foraging
\item lwei - linear weighting model
\item rscpt - risk-sensitive-foraging cpt model
\item rsfft - risk-sensitive foraging fast and frugal tree
\item utility - utility models
}
}
\keyword{internal}
