% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogscimodel-class.R
\docType{data}
\name{Cogscimodel}
\alias{Cogscimodel}
\title{Class for cognitive models}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Cogscimodel$new(formula, data, allowedparm)
}
\arguments{
\item{formula}{Formula (e.g., \code{y ~ x1 + x2}); depends on the model.}

\item{data}{Data.frame or matrix holding \code{formula}'s variables.}

\item{allowedparm}{Matrix. Row names define the names of the allowed parameters, excluding choicerule parameters. Needs 4 columns named \code{"ll","ul","init","na"}, defining each parameter's lower limit, upper limit, starting value for fitting, and value that makes a parameter have zero effect.}

\item{fixed}{(optional) List of model parameters and their fixed values, those parameters won't be estimated. Fixing a parameter to \code{NA} means ignoring it entirely (see details).}

\item{choicerule}{(optional, default \code{NULL}) String specifying the choice rule. Allowed values: \code{NULL} or e.g., \code{"softmax"}, values of \code{type}s in \link[cogsciutils]{choicerule}.}

\item{model}{(optional) String, the model's name.}

\item{discount}{(optional) Integer or integer vector, which or how many many trials to discount}

\item{fit.options}{(optional) List with parameter estimation options:
\describe{
   \item{\code{measure}}{Goodness-of-fit measure (default: \code{"loglikelihood"}); allowed are values of \link[cogsciutils]{gof}'s argument \code{type}}
   \item{\code{n}}{Number of observations, if data is aggregated}
   \item{\code{nbest}}{How many best values to select for grid+hillclimbing?}
   \item{\code{newdata}}{New data to fit parameters with}
   \item{\code{options}}{Other options, see \link[cogsciutils]{gof}}
}}

\item{choicerule}{string holding the choicerule
Remove choicerule
Set data to ignore when fitting}

\item{x}{intteger or integer vector, rows to discount
Get parameter
Set new parameter}

\item{x}{named vector with parameter values, names must be one in \code{allowedparm}
Add a new parameter}

\item{x}{named vector with parameter value not in \code{allowedparm}
Number of free parameter}

\item{x}{String specifying which
Sets fit options
Transforms predictions with choicerule}

\item{x}{string, choice rules, e.g. 'softmax', allowed choicerules see \link[cogsciutils]{choicerule}
Compute goodness of fit}

\item{type}{string, fit measure to use, e.g. 'loglikelihood', allowed types see \link[cogsciutils]{gof}}
}
\description{
Class for cognitive models
}
\details{
Ignore a model parameter by setting it to \code{NA} in \code{fixed}, in this case your matrix \code{allowedparm} needs to contain a value in the column na nullifying the effect of the parameter.
}
\section{Set choice rule}{
NA
}

\examples{
Cogscimodel$new(formula = y ~ x, data = data.frame(y=1:2, x=3:4),
allowedparm = rbind(alpha = c('ul'=0, 'll'=1, 'init'=.5)),
model = 'name of my model')
}
\keyword{datasets}
