% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\docType{package}
\name{cogscimodels-package}
\alias{cogscimodels}
\alias{cogscimodels-package}
\title{cogscimodels: Cognitive Model Estimation and Prediction}
\description{
Implementation of cognitive models, includes a general class cognitive models that has methods to fit free parameters to data and, can apply choicerules, and various goodness-of-fit measures (loglikelihood, MSE, etc.). The interface is similar to the lm() interface. So far an exemplar model "emb" (generalized context model) is included.
}
\details{
you will only need the exemplar-based models [ebm()] function
# Other models in planning
cpt - cummulative prospect theory (Tversky & Kahneman, 1992)
emb - exemplar-based models
hm1988 - Houston & McNamara's dynamic programming model
lwei - linear weighting model
rscpt - risk-sensitive-foraging cpt model
rsfft - risk-sensitive foraging fast and frugal tree
}
\keyword{internal}
