% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm-methods.R
\name{constraints}
\alias{constraints}
\alias{constraints.cm}
\title{Show the constraints of a cognitive model}
\usage{
constraints(x)

\method{constraints}{cm}(x, ...)
}
\arguments{
\item{x}{a model object of class cm}

\item{...}{other arguments (ignored)}
}
\description{
\code{constraints(x)} shows the parameter constraints  of a cognitive model named \code{x} nicely formatted
}
\examples{
D <- data.frame(x = 1, y = 1, z = 1)
M <- bayes_beta(y ~ x + z, D, fix = "start")
constraints(M)    # view the parspace
}
