% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cpttest.R
\docType{data}
\name{cpttest}
\alias{cpttest}
\title{Test data for fitting the Cumulative Prospect Theory, cpt, model}
\format{An object of class \code{"data.table"}; see \code{\link[qtl]{fread}}.}
\source{
Glöckner, A., & Pachur, T. (2012). Cognitive models of risky choice: Parameter stability and predictive accuracy of prospect theory. Cognition, 123(1), 21–32. doi:\href{https://doi.org/10.1016/j.cognition.2011.12.002}{10.1016/j.cognition.2011.12.002}
}
\description{
Data about risky choices to test the parameter estimation method of
the model.
}
\examples{
data(cpttest)
}
\references{
Glöckner, A., & Pachur, T. (2012). Cognitive models of risky choice: Parameter stability and predictive accuracy of prospect theory. Cognition, 123(1), 21–32. doi:\href{https://doi.org/10.1016/j.cognition.2011.12.002}{10.1016/j.cognition.2011.12.002}
}
\keyword{data}
\keyword{datasets}
