% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.combine_constraints}
\alias{.combine_constraints}
\title{Combine multiple constraints of type "constraint" from ROI
Combine constraints, ignore \code{NULL} constraints}
\usage{
.combine_constraints(...)
}
\arguments{
\item{...}{Objects of the class "constraint" from the ROI package (see \link[ROI]{L_constraint})}
}
\value{
An object of class "constraint" with all constraints or \code{NULL} if all \code{...} are \code{NULL}.
}
\description{
Combine multiple constraints of type "constraint" from ROI
Combine constraints, ignore \code{NULL} constraints
}
