% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.grid}
\alias{.grid}
\title{Make a regular or random grid}
\usage{
.grid(rsum, ncol, nrow, regular = TRUE, nstep = NULL, offset = 0)
}
\arguments{
\item{rsum}{row sums}

\item{ncol}{number of columns}

\item{nrow}{number of rows (auto-determines)}

\item{regular}{if regular, \code{FALSE} means random}

\item{nstep}{number of steps, e.g. 4 is 0, 0.25, 0.5, 0.75, 1}

\item{offset}{adjust max and min minus this value}
}
\description{
Make a regular or random grid
}
