% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.simplify_constraints}
\alias{.simplify_constraints}
\title{Simplify constraints}
\usage{
.simplify_constraints(x)
}
\arguments{
\item{x}{An object of type L_constraint from the package ROI}
}
\description{
Simplify a constraint by removing the fully-determined parameter
}
