% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.tabular}
\alias{.tabular}
\title{Makes a tabular for roxygen out of a data frame}
\usage{
.tabular(df, ...)
}
\arguments{
\item{df}{A data.frame

#' @noRd}
}
\description{
This function converts a data frame into a string that can be used 
 to show a table in Roxygen2 documentation
}
