% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ebm_cpp}
\alias{ebm_cpp}
\title{Exemplar-based prediction computation}
\usage{
ebm_cpp(values, features, w, r, q, lambda, lastLearnTrial, firstOutTrial)
}
\arguments{
\item{values}{numeric vector with experienced values}

\item{features}{numeric matrix with feature values}

\item{w}{numeric vector of weights (model parameter)}

\item{r}{square root in distance metic (model parameter)}

\item{q}{exponent in distance metric (model parameter)}

\item{lambda}{sensitivity (model parameter)}

\item{lastLearnTrial}{integer last trial of learning phase}

\item{firstOutTrial}{integer first trial of output, starting the predictions later}
}
\description{
Exemplar-based prediction computation
}
\examples{
# none
}
