% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitivemodel-methods.R
\name{end}
\alias{end}
\title{Ends building a cognitivemodel via \verb{+}}
\usage{
end(obj, ...)
}
\arguments{
\item{obj}{a model object of class 'cognitivemodel'}

\item{...}{other arguments (ignored)}
}
\description{
Ends building a cognitivemodel via \verb{+}
}
