% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitivemodel-methods.R
\name{fit}
\alias{fit}
\title{Estimates free parameters of a cognitvemodel generated via \verb{+}}
\usage{
fit(obj, ...)
}
\arguments{
\item{obj}{a model object of class 'cognitivemodel'}

\item{...}{other arguments (ignored)}
}
\description{
Estimates free parameters of a cognitvemodel generated via \verb{+}
}
