% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitivemodel-methods.R
\name{fun}
\alias{fun}
\title{Adds a function to a cognitivemodel via \verb{+}}
\usage{
fun(obj, ...)
}
\arguments{
\item{obj}{a model object of class 'cognitivemodel'}

\item{...}{other arguments (ignored)}
}
\description{
Adds a function to a cognitivemodel via \verb{+}
}
