% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ev}
\alias{get_ev}
\title{Gets the expected value of gambles}
\usage{
get_ev(x)
}
\arguments{
\item{x}{a risky gamble matrix: x, p, x, p with options in dimension 3}
}
\description{
Gets the expected value of gambles
}
