% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_id_in_grid}
\alias{get_id_in_grid}
\title{Gets one element from a parameter id grid}
\usage{
get_id_in_grid(id, grid)
}
\arguments{
\item{id}{The id to retrieve}

\item{grid}{A list with the parameter grid and ids
resulting from \code{make_grid()}}
}
\description{
Retrieves the id'th value within the parameter id list in grid
}
