% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_p}
\alias{get_p}
\title{Gets the probabilities of gambles}
\usage{
get_p(x)
}
\arguments{
\item{x}{a risky gamble matrix: x, p, x, p with options in dimension 3}
}
\description{
Gets the probabilities of gambles
}
