% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_var}
\alias{get_var}
\title{Gets the variance of gambles}
\usage{
get_var(x)
}
\arguments{
\item{x}{a risky gamble matrix: x, p, x, p with options in dimension 3}
}
\description{
Gets the variance of gambles
}
