% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_x}
\alias{get_x}
\title{Gets the outcomes of gambles}
\usage{
get_x(x)
}
\arguments{
\item{x}{a risky gamble matrix: x, p, x, p with options in dimension 3}
}
\description{
Gets the outcomes of gambles
}
