% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm-methods.R
\name{getCall.cm}
\alias{getCall.cm}
\title{Get the Call to a cognitive model object}
\usage{
\method{getCall}{cm}(x, ...)
}
\arguments{
\item{x}{An existing fit from a cognitive model function such as \link{cm}}
}
\description{
Get the Call to a cognitive model object
}
