% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm-methods.R
\name{gof}
\alias{gof}
\alias{logLik.cm}
\alias{MSE}
\alias{SSE}
\alias{SSE.cm}
\title{Computes Various Model Fit Measures}
\usage{
\method{logLik}{cm}(x, ...)

\code{MSE(x)}

SSE(x)

\code{SSE(x)}
}
\arguments{
\item{x}{a cm object}

\item{...}{other arguments (ignored)}
}
\value{
A number measuring the goodness of fit between predictions and observed data.
}
\description{
\code{logLik(m)} computes the log likelihood of a cm object, \code{SSE(m)} computes the sum of squared errors, \code{MSE(m)} computes the mean squared error.
}
\details{
If a model predicts several values the error measures use the first column of predictions to compute the errors. For example, if the predictions are pr(x) and pr(z), the sum of squared errors is based on the data - pr(x).
}
\examples{
MSE(M)     # 0.1805

D <- data.frame(x = 1, y = 1:1, z = 0:1)
M <- bayes_beta(y ~ x + z, D, fix = "start")
# If you want, look at the predictions
# predict(M)

SSE(M)     # 0.361

}
\concept{{fit measures for cognitive models}}
