% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jr19mem.R
\name{jr19mem}
\alias{jr19mem}
\title{Memory-based preference model (cognitive model)}
\usage{
jr19mem(formula, data)
}
\arguments{
\item{formula}{an object of class formula (\code{response ~ attr1 + attr + attr3 + attr4 | subjvalue | price }). Specifies the variables for response, features, experienced values, and prices; the value variable must be preceeded by a pipe (\code{|}) and the price/cost variable by a second pipe.}

\item{data}{a data.frame or matrix, must contain the variables in \code{formula}; must be ordered in the order in which respondents saw the stimuli.}
}
\value{
An object of class R6 holding the fitted model. Parameters will be fitted by the grid-and-maximum-likelihood method; a model object \code{M} can be viewed with \code{M}, parameter estimates can be obtained with \code{mod$param}
}
\description{
Fits the memory-based model used in Jarecki & Rieskamp (2019). The interface is similar to lm.
}
\details{
The function is a wrapper around the more general \code{ebm} exemplar-based-model function.
}
\examples{
# Create data
attributes <- expand.grid(a1=0:1,a2=0:1)[rep(1:4,2),]
prices <- rep(c(0,.5,.5,1),2)
subj_values <- rep(c(2,0.2,0.1,2), 2)
choices <- rep(c(1,0,0,1),2)
dt <- data.frame(choices, attributes, prices, subj_values)

# Fit the model
M <- jr19mem(choices ~ a1 + a2 | subj_values | prices, data = dt)

#' M # view results
M$predict() # predict data
M$coef() # view coefficients

}
\author{
Jana B. Jarecki, \email{jj@janajarecki.com}
}
