% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwei.R
\name{lwei}
\alias{lwei}
\title{Linear-weighting model}
\usage{
lwei(formula, nopt, nout, ref, fix = NULL, data, choicerule,
  weighting = c("TK1992"), value = c("TK1992"))
}
\description{
Linear-weighting model
}
\keyword{internal}
