% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_grid_id_list}
\alias{make_grid_id_list}
\title{Make a grid for parameters}
\usage{
make_grid_id_list(names, nsteps = list(), ub = list(), lb = list(),
  sumto = NULL, to = 1, offset = 0, ...)
}
\arguments{
\item{names}{The parameter names}

\item{nsteps}{The number of steps for each parameter in \code{names}}

\item{ub}{A named list, upper bounds for the parameters in \code{names}}

\item{lb}{A named list, lower bounds for the parameters in \code{names}}

\item{sumto}{(optional) A named list with vectors of parameter names of
parameters that should be constrained to sum to a value}

\item{to}{(optional) A list, values that parameters should sum to,
the names must equal the names of \code{sumto}}

\item{offset}{(optional) A small numeric value to offset the grid from the
bounds of the parameter values}
}
\description{
Make a list with a regularly-spaced parameter grid for grid-based fitting
 and parameter recovery
}
