% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm-methods.R
\name{nobs}
\alias{nobs}
\alias{nobs.cm}
\title{Number of Parameters, Attributes, and Stimuli}
\usage{
nobs(x, ...)

nobs(x)
}
\arguments{
\item{x}{a model object}
}
\description{
Number of Parameters, Attributes, and Stimuli
}
\examples{
nobs(M)
}
