% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm-methods.R
\name{npar}
\alias{npar}
\alias{npar.cogscimodel}
\alias{nobs.cogscimodel}
\alias{nstim}
\alias{natt}
\title{Number of Parameters, Attributes, Stimuli}
\usage{
\code{npar(x)}

\method{npar}{cogscimodel}(x, type = "all")

\code{nobs(x)}

\code{nstim(x)}

\code{natt(x)}
}
\arguments{
\item{x}{a model object}

\item{...}{other arguments (ignored)}

\item{type}{A string: \code{"all"} counts all parameters, \code{"free"} counts the free parameters, \code{"fix"} countes the fixed parameters, \code{"constrained"} counts constrained parameters.}
}
\description{

}
\examples{
D <- data.frame(x = 1, y = 1, z = 1)
M <- bayes_beta(y ~ x + z, D, fix = "start")
npar(M) # 3

# Print the model
M 

# Bayesian model
# Call:  
# bayes_beta(formula = y ~ x + z, data = D, fix = "start")  
#
# Constrained parameter:
# delta      x      z
#     1      1      1


nobs(M)
nstim(M)
natt(M)
}
