% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm-methods.R
\name{nstim}
\alias{nstim}
\title{Number of Parameters, Attributes, and Stimuli

@name nstim}
\usage{
nstim(x)
}
\arguments{
\item{x}{a model object}
}
\description{
Number of Parameters, Attributes, and Stimuli

@name nstim
}
\examples{
nstim(M)
}
