% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm-methods.R
\name{parspace}
\alias{parspace}
\alias{parspace.cm}
\title{Show the paramter space of a cognitive model}
\usage{
parspace(x)

\method{parspace}{cm}(x, ...)
}
\arguments{
\item{x}{a model object of class cm}

\item{...}{other arguments (ignored)}
}
\description{
\code{parspace(cm)} shows the parameter names and upper and lower bounds of parameters in a cognitice model \code{cm}
}
\examples{
D <- data.frame(x = 1, y = 1, z = 1)
M <- bayes_beta(y ~ x + z, D, fix = "start")
parspace(M)    # view the parspace
}
