% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm-methods.R
\name{predict.cm}
\alias{predict.cm}
\title{Predictions from Cognitive Models (class cm)}
\usage{
\method{predict}{cm}(object, ..., type = "response", newdata = NULL)
}
\arguments{
\item{object}{A cognitive model object of class 'cm'}

\item{...}{Additional cognitive models}

\item{type}{A string, what to predict, usually \code{"response"}, but other values may be possible, see the help pages of the model you use.}

\item{newdata}{A data.frame with new data to predict}
}
\value{
A vector or matrix of predictions, if multiple models are supplied using \code{...}, returns a list containing the predictions for each model
}
\description{
Predictions from Cognitive Models (class cm)
}
