% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cogscimodel.R
\name{predict.cogscimodel}
\alias{predict.cogscimodel}
\title{Predictions from Cognitive Models}
\usage{
\method{predict}{cogscimodel}(obj, ..., type = "response", newdata = NULL)
}
\arguments{
\item{obj}{A cognitive models with class cogscimodel}

\item{...}{Additional cognitive models}

\item{type}{A string specifying what to predict, usually it is the \code{"response"}, but other values may be possible, see the help pages of the model being used}

\item{newdata}{A data.frame with new data to predict}
}
\value{
A vector or matrix of predictions, if multiple models are supplied using \code{...}, returns a list containing the predictions for each model
}
\description{
Predictions from Cognitive Models
}
