% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-constraints.R
\name{print.csm_constraint}
\alias{print.csm_constraint}
\title{Prints the constraints of a cogscimodel object nicely}
\usage{
\method{print}{csm_constraint}(x, ..., latex = FALSE)
}
\arguments{
\item{latex}{(optional) if \code{TRUE} formats them for LaTeX documents}
}
\description{
or NULL if there are no constraints
}
