% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover_par-class.R
\name{recover_par}
\alias{recover_par}
\title{Class for parameter recovery objects}
\usage{
recover_par(model, nruns = 100, nsteps = list(), seed = 98634)
}
\arguments{
\item{model}{An object of class cm.}

\item{nruns}{(optional, default 100) Integer, how often to repeat the recovery per set of true parameter?}

\item{nsteps}{(optional, default 3) list with steps by which to vary each parameter.}

\item{seed}{(optional) Random seed for reproducibility.}
}
\description{
Parameter recovery for a model works as follows: First, select a set of true model parameters \eqn{\theta}. Second, simulate data from a \link{Cognitivemodel} with the true parameters \eqn{\theta}. Third, use the simulated data to estimate the cognitive model's free parameters, store the best-fitting parameter estimates \eqn{\theta*}. Repeat step 1-2 a number of runs Lastly, diagnose the results if the recovered parameter equal the true parameter.
}
\details{
Conducts a parameter recovery from  cognitive model objects (\link{Cm} objects).
}
\examples{
No examples yet
}
\author{
Jana B. Jarecki, \email{jj@janajarecki.com}
}
