% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscpt.R
\name{rscpt}
\alias{rscpt}
\title{Risk-sensitive foraging cognitive prospect theory
#' @inheritParams Cogscimodel rscpt}
\usage{
rscpt(env, formula = NULL, fixed = NULL, data = NULL, choicerule,
  ref)
}
\description{
Risk-sensitive foraging cognitive prospect theory
#' @inheritParams Cogscimodel rscpt
}
