% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsfft.R
\name{rsfft}
\alias{rsfft}
\title{Risk-sensitive foraging fast and frugal tree}
\usage{
rsfft(formula = NULL, sbt = NULL, nopt = NULL, nout = NULL,
  fixed = NULL, data = NULL, env = NULL, choicerule,
  terminal.fitness.fun)
}
\arguments{
\item{formula}{Formula, like \code{y ~ x1 + x2}, depends on the model.}

\item{fixed}{(optional) Named vector with model parameter(s) that are predefined (will not be fitted). If a fixed parameter is \code{NULL} model parameter will be ignored. See details.}

\item{data}{Data.frame or matrix, must contain all variables in \code{formula}.}

\item{choicerule}{(optional, defaultoptions \code{NULL}) String specifying the choice rule, e.g. \code{'softmax', 'argmax'}, see \link[cogsciutils]{choicerule} for possible choicerules; \code{NLL} means predictions as is.}
}
\description{
Risk-sensitive foraging fast and frugal tree
}
