% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shortfall_cpp}
\alias{shortfall_cpp}
\title{Exemplar-based prediction computation}
\usage{
shortfall_cpp(x, p, a, beta, delta)
}
\arguments{
\item{x}{numeric matrix with outcomes}

\item{p}{numeric matrix with probabilities}

\item{beta}{parameter}

\item{delta}{parameter
'}

\item{w}{numeric vector with aspiration leel}
}
\description{
Exemplar-based prediction computation
}
