% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shortfalltest.R
\docType{data}
\name{shortfalltest}
\alias{shortfalltest}
\title{Test data for fitting the Shortfall model}
\format{An object of class \code{"data.table"}; see \code{\link[qtl]{fread}}.}
\source{
Andraszewicz, S. (2014). Quantitative analysis
of risky decision making in economic environments (Doctoral dissertation,
University_of_Basel). \href{https://edoc.unibas.ch/33427/}{doi:10.5451/unibas-006268585}
}
\description{
Data about risky choices to test the parameter estimation method of
the shortfall model.
}
\examples{
data(shortfalltest)
}
\references{
Andraszewicz, S. (2014). Quantitative analysis
of risky decision making in economic environments (Doctoral dissertation,
University_of_Basel). \href{https://edoc.unibas.ch/33427/}{doi:10.5451/unibas-006268585}
}
\keyword{data}
\keyword{datasets}
