% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain_models.R
\name{start}
\alias{start}
\title{Start chain of cogscimodels}
\usage{
start(data, ...)
}
\arguments{
\item{data}{a data.frame holding the variables that the first model uses}
}
\description{
Start chain of cogscimodels
}
