% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tversky1992ex.R
\docType{data}
\name{tversky1992ex}
\alias{tversky1992ex}
\title{Example data with two risky gambles}
\format{An object of class \code{data.frame}}
\source{
Tversky, A., & Kahneman, D. (1992). Advances in prospect theory: Cumulative representation of uncertainty. \emph{Journal of Risk and Uncertainty, 5(4)}, 297–323. doi\href{https://doi.org/10.1007/BF00122574}{10.1007/BF00122574}
}
\description{
Data with two example risky choices with two gambles
}
\examples{
data(tversky1992ex)
}
\references{
Tversky, A., & Kahneman, D. (1992). Advances in prospect theory: Cumulative representation of uncertainty. \emph{Journal of Risk and Uncertainty, 5(4)}, 297–323. doi\href{https://doi.org/10.1007/BF00122574}{10.1007/BF00122574}
}
\keyword{data}
\keyword{datasets}
