% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-baseline.R
\name{baseline_const_c}
\alias{baseline_const_c}
\alias{baseline_const_d}
\alias{baseline_mean_c}
\alias{baseline_mean_d}
\title{Baseline models}
\usage{
baseline_const_c(formula, const, data, ...)

baseline_const_d(formula, const, data, ...)

baseline_mean_c(formula, data, ...)

baseline_mean_d(formula, data, ...)
}
\arguments{
\item{formula}{A formula, the observed variables, with only a left-hand-side (for example: \code{response ~ .}), note the "\code{~ .}" on the right.}

\item{const}{A number, the value that the model predicts  (only in \code{baseline_const_c/d}) .}

\item{data}{A data frame.}
}
\value{
Returns a cognitive model object, which has the class \code{cm}(cm). A model, that is saved as \code{m}, can be summarized with \code{summary(m)} or \code{anova(m)}. The parameter space can be viewed using \code{parspace(m)}, constraints can be viewed using \code{constraints(m)}.
}
\description{
Fits baseline models. Baseline models are stimulus-agostic models used as sanity checks in cognitive model comparisons. Other cognitive models should beat a baseline model -- if not, the other cognitive models don't describe patterns in the responses well.
\itemize{
\item \code{baseline_const_c()} predicts a constant value for continuous responses.
\item \code{baseline_const_d()} predicts a constant value for discrete responses.
\item \code{baseline_mean_c()} fits the mean of the observed responses for continuous responses.
\item \code{baseline_mean_d()} fits the mean of the observed responses for discrete responses.
}
}
\details{
Use \strong{\code{baseline_xxx_d}} for discrete response variables; and  \strong{\code{baseline_xxx_c}} for continuous response variables.

\code{baseline_const_} predicts the value of \code{const} for all trials. For example \code{const = 0.50} is a common baseline for two-option choice tasks, and it predicts \emph{Pr=0.50} for each trial. The model \code{baseline_mean_} predicts the mean of the observed responses for each trial. This model has one free parameter (\code{m}, the mean).
}
\examples{
# Here is some data, let y be the observed data
D <- data.frame(y = c(1,1,0), x = c(1,2,3))

# Baseline model that predicrs Pr = 0.50
M <- baseline_const_d(y ~ ., const = 0.50, data = D)

predict(M)                         # predicts 0.5, 0.5, 0.5
npar(M)                            # 0 parameter
logLik(M)                          # log likelihood (binomial)

M <- baseline_mean_d(y ~ ., D)     # Pr = mean(observed variable)
predict(M)                         # predicts 0.66, 0.66, 0.66
coef(M)                            # mean counts as free parameter
npar(M)                            # 1 free parameter, the mean

}
\seealso{
Other cognitive models: 
\code{\link{Cm}}
}
\concept{cognitive models}
