% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-formulas.R
\name{chr_as_rhs}
\alias{chr_as_rhs}
\title{Convert a character to formula}
\usage{
chr_as_rhs(x)
}
\arguments{
\item{x}{A string or RHS formula}
}
\description{
Checks if x is a character and if so converts it into RHS formula
}
