% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm-class.R
\name{Cogscimodel}
\alias{Cogscimodel}
\title{The R6 class for Cognitive Model Objects}
\description{
The R6 class for Cognitive Model Objects

The R6 class for Cognitive Model Objects
}
\details{
\code{Cogscimodel$new(formula, data, parspace)}

\code{parspace}. It is optional to define a value that makes the parameter have zero effect in the column called "na" in \code{parspace}. For example, a parameter \code{b} in \code{b*x}, has no-effect  when setting \code{b=0}. The no-effect value can be \code{NA}.

\bold{fix}
 Supplying \code{list(alpha=0.5, beta="alpha", gamma=NA)} constrains the parameters alpha, beta, and gamma as follows
\itemize{
\item{\code{alpha=0.5}}{: fix alpha to 0.5.}
\item{\code{beta="alpha"}}{: fix beta to the value of alpha; alpha may be a free or fixed model parameter in \code{parspace}.}
\item{\code{delta=NA}}{: ignore delta if it can be ignored by setting delta equal to the value in the column "na" of \code{parspace}, given that parspace has a value in the column "na" for delta.}
}
You can ignore a model parameter by setting it to \code{NA} in \code{fix}, in this case your \code{parspace} needs to contain a value in the column na nullifying the effect of the parameter.
}
\examples{
Cogscimodel$new(
    formula = y ~ x,
    data = data.frame(y=1:2, x=3:4),
    parspace = rbind(alpha = c("ub"=0, "lb"=1, "start"=.5)),
    title = 'name of my model'))
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{title}}{A string, the name of the model}

\item{\code{formula}}{Calling formula, e.g., \code{y ~ x1 + x2 + x3} models one response variable (\code{y}) given one three-attribute stimulus, and \code{y ~ x1 + x2 | z1 + z2} models a response given two two-attribute stimuli where attributes are separated by a "\code{+}"" and stimuli are separated by a "\code{|}".}

\item{\code{input}}{Inputs to the model in a standardized form: three-dimensional matrix based on \code{formula} and \code{data} with dimensions n(data) x n(attributes) x n(stimuli)}

\item{\code{more_input}}{Additional inputs to the model from variables that are not part of the formula, three-dimensional matrix with dimensions n(data) x n(additional inputs) x n(stimuli)}

\item{\code{res}}{The response variable to be modeled, two-dimensional matrix with dimensions n(data) x n(response variables)}

\item{\code{natt}}{List with the number of attributes per stimulus}

\item{\code{nobs}}{Number of observations in \code{data}}

\item{\code{nstim}}{Number of stimuli in model input}

\item{\code{nres}}{Number of response variables}

\item{\code{ncon}}{Number of constraints on the parameter}

\item{\code{par}}{Model parameters, a named list}

\item{\code{parspace}}{Parameter space, a matrix with one parameter per row, and four columns lb (lower bound), ub (upper bound), start (starting value for fitting) and na (value of the parameter if it has no effect on the model); rownames are parameter names}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Cogscimodel$new()}}
\item \href{#method-fit}{\code{Cogscimodel$fit()}}
\item \href{#method-coef}{\code{Cogscimodel$coef()}}
\item \href{#method-predict}{\code{Cogscimodel$predict()}}
\item \href{#method-simulate}{\code{Cogscimodel$simulate()}}
\item \href{#method-set_data}{\code{Cogscimodel$set_data()}}
\item \href{#method-set_par}{\code{Cogscimodel$set_par()}}
\item \href{#method-get_par}{\code{Cogscimodel$get_par()}}
\item \href{#method-npar}{\code{Cogscimodel$npar()}}
\item \href{#method-gof}{\code{Cogscimodel$gof()}}
\item \href{#method-logLik}{\code{Cogscimodel$logLik()}}
\item \href{#method-BIC}{\code{Cogscimodel$BIC()}}
\item \href{#method-AIC}{\code{Cogscimodel$AIC()}}
\item \href{#method-AICc}{\code{Cogscimodel$AICc()}}
\item \href{#method-MSE}{\code{Cogscimodel$MSE()}}
\item \href{#method-SSE}{\code{Cogscimodel$SSE()}}
\item \href{#method-RMSE}{\code{Cogscimodel$RMSE()}}
\item \href{#method-set_lb}{\code{Cogscimodel$set_lb()}}
\item \href{#method-set_ub}{\code{Cogscimodel$set_ub()}}
\item \href{#method-set_start}{\code{Cogscimodel$set_start()}}
\item \href{#method-print}{\code{Cogscimodel$print()}}
\item \href{#method-summary}{\code{Cogscimodel$summary()}}
\item \href{#method-clone}{\code{Cogscimodel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cogscimodel$new(
  formula,
  data = NULL,
  parspace = make_parspace(),
  fix = NULL,
  choicerule = NULL,
  title = NULL,
  discount = NULL,
  mode = NULL,
  options = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formula}}{A formula (e.g., \code{y ~ x1 + x2}).}

\item{\code{data}}{A data frame containing \code{formula}'s variables.}

\item{\code{parspace}}{(optional, \bold{required} to add model parameters) A amtrix with the parameter space: a \code{n x 4} matrix with columns named \code{"lb","ub","start","na"}. Easy to create using \link{make_parspace}, e.g. \code{make_parspace(alpha=c(0,1))}. The matrix' row names are parameter names, columns contain the lower limit, upper limit, starting value in fitting, and (optional) a value that makes a parameter have zero effect, which can be NA. See details.}

\item{\code{fix}}{(optional) Parameter constraints. Can be \code{"start"} or a list with \code{parname=value}-pairs. Parameter names see below under "parameter space".
\itemize{
\item{\code{"start"} constrains all available parameters to their starting values. Useful for model testing.}
\item{\code{parname = 0.5} constrains a parameter to 0.5.}
\item{\code{parname = "p2"} constrains a parameter to another model parameter \code{p2}.}
\item{\code{parname = NA} tells the model to omit a parameter, if possible.}
}}

\item{\code{choicerule}}{(optional) String specifying the choice rule (default \code{NULL}), allowed are \code{NULL}, \code{"softmax"}, \code{"eps"}, and other values of the argument \code{type} to \link[cogsciutils]{choicerule}.}

\item{\code{title}}{(optional, default is the class name) A string, the model's name.}

\item{\code{discount}}{(optional) An integer or integer vector (default \code{0}), ddefining which or how many, starting from trial 1, to discount when fitting.}

\item{\code{mode}}{A string. Allowed are \code{"discrete"}, \code{"continuous"}, specifies the model's response mode. Discrete responses are binary (0 or 1), continuous responses are numbers with a normal error.}

\item{\code{options}}{(optional) Options to control the parameter fitting methods, see the "Options" section of \code{\link{cogscimodel}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\subsection{Method \code{fit()}}{
Fits the free model parameters to the response data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cogscimodel$fit(
  solver = self$options$fit_solver,
  measure = self$options$fit_measure,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{solver}}{(optional) A string with the model solver}

\item{\code{measure}}{(optional) A string with the goodness-of-fit measure that the solver optimizes (e.g. \code{"loglikelihood"}). Possible values, see the \code{type} argument in \link[cogsciutils]{gof}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-coef"></a>}}
\subsection{Method \code{coef()}}{
Returns the free parameters of the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cogscimodel$coef()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\subsection{Method \code{predict()}}{
Make predictions for every row in the input data to the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cogscimodel$predict(type = "response", newdata = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(optional, default \code{"response"}) Type of prediction to make. See the respective model.}

\item{\code{newdata}}{(optional) A data frame with new data for which to compute predictions.
calls the child model's make_prediction function for
every slice of the third dimension of the input matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-simulate"></a>}}
\subsection{Method \code{simulate()}}{
Simulates observable data from the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cogscimodel$simulate(newdata = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newdata}}{(optional) A data frame with inputs (e.g., stimuli) to be used in the simulation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A matrix with simulated choices or judgment values
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_data"></a>}}
\subsection{Method \code{set_data()}}{
New data input for a cogscim
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cogscimodel$set_data(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A data frame with variables corresponding to the inputs that the model needs}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_par"></a>}}
\subsection{Method \code{set_par()}}{
Set model parameter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cogscimodel$set_par(x, check = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A named list of parameters, their names must be in \code{parspace}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_par"></a>}}
\subsection{Method \code{get_par()}}{
Get the model parameter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cogscimodel$get_par(x = "all")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-npar"></a>}}
\subsection{Method \code{npar()}}{
Number of model parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cogscimodel$npar(x = "all")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{String specifying which of the parameters to return, allowed are \code{"all", "free", "constrained", "equal"}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gof"></a>}}
\subsection{Method \code{gof()}}{
Computes the goodness of model fit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cogscimodel$gof(
  type,
  n = self$options$fit_n,
  newdata = self$options$fit_data,
  discount = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{A string, fit measure to use, e.g. \code{"loglikelihood"}, for the allowed types see \link[cogsciutils]{gof}}

\item{\code{discount}}{(optional) An integer or integer vector (default \code{0}), ddefining which or how many, starting from trial 1, to discount when fitting.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logLik"></a>}}
\subsection{Method \code{logLik()}}{
Log likelihood of the observed responses under the model predictions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cogscimodel$logLik(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BIC"></a>}}
\subsection{Method \code{BIC()}}{
Bayesian Information Criterion of the model predictions given the observed responses
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cogscimodel$BIC(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AIC"></a>}}
\subsection{Method \code{AIC()}}{
Akaike Information Criterion of the model predictions given the observed response
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cogscimodel$AIC(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AICc"></a>}}
\subsection{Method \code{AICc()}}{
Small-sample corrected Akaike Information Criterion of the model predictions given the responses
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cogscimodel$AICc(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MSE"></a>}}
\subsection{Method \code{MSE()}}{
Mean squared error between the model predictions given the observed responses
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cogscimodel$MSE(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SSE"></a>}}
\subsection{Method \code{SSE()}}{
Sum of squared errors between the model predictions given the observed responses
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cogscimodel$SSE(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RMSE"></a>}}
\subsection{Method \code{RMSE()}}{
Root mean squared error between the model predictions and the observed responses
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cogscimodel$RMSE(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_lb"></a>}}
\subsection{Method \code{set_lb()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cogscimodel$set_lb(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{new parameter bound, e.g. \code{alpha = 0} sets the lower bound of the parameter \code{alpha} to 0}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_ub"></a>}}
\subsection{Method \code{set_ub()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cogscimodel$set_ub(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{new parameter bound, e.g. \code{alpha = 0} sets the lower bound of the parameter \code{alpha} to 0}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_start"></a>}}
\subsection{Method \code{set_start()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cogscimodel$set_start(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{new parameter bound, e.g. \code{alpha = 0} sets the lower bound of the parameter \code{alpha} to 0}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Prints the model object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cogscimodel$print(digits = 2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{digits}}{A number specifying the number of digits to print}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\subsection{Method \code{summary()}}{
Summarizes the model, it's parameters, and some fit information
Transforms predictions with choicerule
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cogscimodel$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cogscimodel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
