% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogscimodel-class.R
\docType{data}
\name{Cogscimodel}
\alias{Cogscimodel}
\title{Class for cognitive models}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Cogscimodel$new(formula, data, allowedparm)
}
\arguments{
\item{formula}{Formula, like \code{y ~ x1 + x2}, depends on the model.}

\item{data}{Data.frame or matrix, must contain all variables in \code{formula}.}

\item{allowedparm}{Matrix with 4 columns like \code{rbind(alpha=c('ul'=0,'ll'=1,'init'=0.5,'na'=1))}. Defines each parameter's lower limit, upper limit, starting value (for fitting), and value that make the parameter have zero effect. Row names define parameter names. Columns names must be \code{'ll', 'ul', 'init', 'na'} (lower limit, upper limit, inital value, no-effect value; resp.).}

\item{fixed}{(optional) Named vector with model parameter(s) that are predefined (will not be fitted). If a fixed parameter is \code{NULL} model parameter will be ignored. See details.}

\item{choicerule}{(optional, default \code{NULL}) String specifying the choice rule, e.g. \code{'softmax', 'argmax'}, see \link[cogsciutils]{choicerule} for possible choicerules; \code{NLL} means predictions as is.}

\item{model}{(optional) String, the name of the model}

\item{discount}{(optional) Integer or integer vector, which or how many many trials to discount?}

\item{x}{intteger or integer vector, rows to discount
Get parameter
Set new parameter}

\item{x}{named vector with parameter values, names must be one in \code{allowedparm}
Add a new parameter}

\item{x}{named vector with parameter value not in \code{allowedparm}
Sets fit options
Transforms predictions with choicerule}

\item{x}{string, choice rules, e.g. 'softmax', allowed choicerules see \link[cogsciutils]{choicerule}
Compute goodness of fit}

\item{type}{string, fit measure to use, e.g. 'loglikelihood', allowed types see \link[cogsciutils]{gof}}
}
\description{
Class for cognitive models
}
\details{
Ignore a model parameter by setting it to \code{NULL}, in this case your matrix \code{allowedparm} needs to contain
}
\section{Set choice rule}{
NA
}

\examples{
Cogscimodel$new(formula = y ~ x, data = data.frame(y=1:2, x=3:4),
allowedparm = rbind(alpha = c('ul'=0, 'll'=1, 'init'=.5)),
model = 'name of my model')
}
\keyword{datasets}
