\name{cogscimodels-package}
\alias{cogscimodels-package}
\alias{cogscimodels}
\docType{package}
\title{
\packageTitle{cogscimodels}
}
\description{
\packageDescription{cogscimodels}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cogscimodels}
\packageIndices{cogscimodels}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{cogscimodels}

Maintainer: \packageMaintainer{cogscimodels}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
