% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-cpt.R
\name{cpt}
\alias{cpt}
\alias{cpt_mem}
\title{Cumulative Prospect Theory Models}
\usage{
cpt(
  formula,
  data,
  choicerule,
  ref = 0L,
  fix = list(),
  weighting = c("TK1992"),
  value = c("TK1992"),
  options = NULL
)

cpt_mem(formula, mem, data, choicerule, editing = "hedonic", ...)
}
\arguments{
\item{formula}{A \link{formula} such as \code{y ~ x1 + p1 + x2 | y1 + py + y2} specifying the columns in \code{data} that contain outcomes, probabilities and (optional) the observations. Lines (\code{|}) separate different gambles. The formula must alternate outcomes and probabilities (x + p + x2 + p2), the last probability can be omitted.}

\item{ref}{(optional, default: 0) A number, string, or RHS \link{formula}. The reference point. Formula and string set the reference point to variables in data, and \code{ref = "myvar"} means \code{data$myvar}, and so does \code{ref = ~ myvar}.}

\item{weighting}{(optional) A string, the weighting function. Currently only \code{"KT1992"}, the value function in Kahneman & Tversky (1992), is implemented.}

\item{value}{(optional) A string, the value function. Currently, only \code{"KT1992"} the value function in Kahneman & Tversky (1992) is implemented.}

\item{mem}{(optional, default: 0) A number, string, or RHS \link{formula}. Thr prior gains or losses in memory. Formula and string refer to variables in the data, and \code{mem = "myvar"}means \code{data$myvar} and so does \code{mem = ~myvar}.}

\item{editing}{(optional, default: \code{"hedonic"}) A string, the editing rule to use (see Thaler & Johnson, 1999, pp. 645). Currently only \code{"hedonic"}.}
}
\value{
An object of class R6 holding the model, it has free parameters. A model object \code{M} can be viewed with \code{M}, choice or utility predictions can be made by \code{predict(M)}, also for new data by \verb{predict(M, newdata = ...}).
}
\description{
cpt_* fit a cumulative prospect theory.
\itemize{
\item \code{cpt()} fits cumulative prospect theory (Tversky & Kahneman, 1992)
\item \code{cpt_mem()} fits cumulative prospect theory with memory (Thaler & Johnson, 1990)
}
}
\details{
Fits cumulative prospect theory.
}
\section{Functions}{
\itemize{
\item \code{cpt_mem}: Prospect Theory with Memory
}}

\examples{
## From Tversky, A., & Kahneman, D. (1992).
dt <- data.frame(
  x1 = c(100, -100),
  px = 1,
  x2 = 0,
  y1 = c(200, -200),
  py = c(.71,.64),
  y2 = 0,
  rp = 1)

# Make the model, add fix parameters (don"t fit)
# using the Parameter from the paper
M <- cpt(rp ~ x1 + px + x2 | y1 + py + y2, ref = 0,
         choicerule = "softmax", data = dt,
         fix = list(alpha = 0.88, beta = 0.88, lambda = 2.25,
         gammap = 0.61, gamman = 0.69, tau = 1))

# View the model
M

# Methods
predict(M, "value") # predict values, also: M$predict("value")
predict(M, "mode") # predict choice probability after softmax
summary(M)
anova(M)
}
\references{
Tversky, A., & Kahneman, D. (1992). Advances in prospect theory: cumulative representation of uncertainty. Journal of Risk and Uncertainty, 5, 297–-323. doi:\href{https://doi.org/10.1007/BF00122574}{10.1007/BF00122574}

Thaler, R. H., & Johnson, E. J. (1990). Gambling with the House Money and Trying to Break Even: The Effects of Prior Outcomes on Risky Choice. Management Science, 36(6), 643–-660. doi:\href{https://doi.org/10.1287/mnsc.36.6.643}{10.1287/mnsc.36.6.643}
}
\author{
Jana B. Jarecki, \email{jj@janajarecki.com}
}
