% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-cpt.R
\name{cpt}
\alias{cpt}
\title{Kahneman & Tversky's (1992) cumulative prospect theory model}
\usage{
cpt(
  formula,
  data = NULL,
  ref = 0L,
  fix = list(),
  choicerule = NULL,
  weighting = c("TK1992"),
  value = c("TK1992"),
  options = NULL
)
}
\arguments{
\item{formula}{A \link{formula} such as \code{y ~ x1 + p1 + x2 | y1 + py + y2} specifying the columns in \code{data} that contain outcomes, probabilities and (optional) the observations. Lines (\code{|}) separate different gambles. The formula must alternate outcomes and probabilities (x + p + x2 + p2), the last probability can be omitted.}

\item{choicerule}{(default \code{"softmax"}) A string specifying the choierule of the model, for instance \code{"softmax"}. If \code{"none"}, the model predicts the utility value of the option. Allowed are the values of the \code{type} argument in \code{\link[cognitiveutils]{choicerule}}.}

\item{weighting}{(optional) weighting function. Currently the one used in Kahneman & Tversky (1992), \code{"KT1992"}, is possible.}

\item{value}{(optional) value function. Currently, only the one used by Kahneman & Tversky (1992), \code{"KT1992"}, is possible.}
}
\value{
An object of class R6 holding the model, it has free parameters. A model object \code{M} can be viewed with \code{M}, predictions can be made with \code{M$predict()} for choice predictions, and \code{M$predict("ev")} for the expected value of the optimal choice and \code{M$predict("value", 1:2)} for the expected value of all choices.
}
\description{
Fits the cumulative prospect theory model.
}
\details{
Fits cumulative prospect theory.
}
\examples{
## From Tversky, A., & Kahneman, D. (1992).
dt <- data.frame(
  x1 = c(100, -100),
  px = 1,
  x2 = 0,
  y1 = c(200, -200),
  py = c(.71,.64),
  y2 = 0,
  rp = 1)

# Make the model, add fix parameters (don't fit)
# using the Parameter from the paper
M <- cpt(rp ~ x1 + px + x2 | y1 + py + y2, ref = 0,
         choicerule = "softmax", data = dt,
         fix = list(alpha = 0.88, beta = 0.88, lambda = 2.25,
         gammap = 0.61, gamman = 0.69, tau = 1))

# View the model
M

# Methods
predict(M, "value") # predict values, also: M$predict("value")
predict(M, "mode") # predict choice probability after softmax
summary(M)
anova(M)
}
\references{
Tversky, A., & Kahneman, D. (1992). Advances in prospect theory: cumulative representation of uncertainty. Journal of Risk and Uncertainty, 5, 297–323. doi:10.1007/BF00122574
}
\author{
Jana B. Jarecki, \email{jj@janajarecki.com}
}
