% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-cpt.R
\name{cpt}
\alias{cpt}
\alias{cpt_d}
\alias{cpt_c}
\alias{cpt_mem_d}
\alias{cpt_mem_c}
\title{Cumulative Prospect Theory Models}
\usage{
cpt_d(
  formula,
  data,
  choicerule,
  ref = 0L,
  fix = list(),
  weighting = c("TK1992"),
  value = c("TK1992"),
  options = NULL
)

cpt_c(
  formula,
  data,
  ref = 0L,
  fix = list(),
  weighting = c("TK1992"),
  value = c("TK1992"),
  options = NULL
)

cpt_mem_d(
  formula,
  mem,
  data,
  fix = list(),
  choicerule,
  editing = "hedonic",
  weighting = c("TK1992"),
  value = c("TK1992"),
  options = NULL
)

cpt_mem_c(
  formula,
  mem,
  data,
  fix = list(),
  editing = "hedonic",
  weighting = c("TK1992"),
  value = c("TK1992"),
  options = NULL
)
}
\arguments{
\item{formula}{A \link[stats:formula]{formula}, the variables in \code{data} to be modeled. For example, \code{y ~ x1 + p2 + x3 + p4 | x5 + p6 + x7 + p8} models response y as function of two stimuli with features x1, p2, x3, p4 and x5, p6, x7, p8, alternating outcomes \code{x} and probabilities \code{p} (respectively). Lines \code{|} separate stimuli.}

\item{data}{A data frame, the data to be modeled.}

\item{choicerule}{A string, the choice rule. Allowed values, see  \code{cm_choicerules()}: \code{"none"} is no choice rule, \code{"softmax"} is \href{https://en.wikipedia.org/wiki/Softmax_function}{soft-maximum}, \code{"luce"} is \href{https://en.wikipedia.org/wiki/Luce\%27s_choice_axiom}{Luce's axiom}.}

\item{ref}{(optional, default: 0) A number, string, or RHS \link[stats:formula]{formula}, the reference point or the variable in \code{data} that holds the reference point. For example \code{~ ref}.}

\item{fix}{(optional) A list with parameter-value pairs of fixed parameters. If missing all free parameters are estimated. If set to \code{"start"} all parameters are fixed to their start values. Model parameter names are \emph{\code{alpha}, \code{beta}, \code{gammap}, \code{gamman}, \code{lambda}} (see details - model parameters).
\itemize{
\item \code{list(alpha = 1.08)} sets parameter \emph{\code{alpha}} equal to 1.08.
\item \code{list(alpha = "beta")} sets parameter \emph{\code{alpha}} equal to parameter \emph{\code{beta}} (estimates \emph{\code{beta}}).
\item \code{list(beta = "alpha", alpha = 1.08)} sets parameter \emph{\code{beta}} equal to parameter \emph{\code{alpha}} and sets \emph{\code{alpha}} equal to 1.08 (estimates none of the two).
\item \code{list(alpha = NA)} omits the parameter \emph{\code{alpha}}, if possible.
\item \code{"start"} sets all parameters equal to their initial values (estimates none). Useful for building a first test model.
}}

\item{weighting}{(optional) A string, name of the probability weighting function, allowed are \code{"KT1992"} for the weighting in Kahneman & Tversky (1992) and \code{NA} for no weighting.}

\item{value}{(optional) A string, the name of the value function. Allowed is only \code{"KT1992"} for the value function in Kahneman & Tversky (1992) and \code{NA} for no value transformation.}

\item{options}{(optional) A list, list entries change the modeling procedure. For example, \code{list(lb = c(k=0))} changes the lower bound of parameter \emph{k} to 0, or \code{list(fit_measure = "mse")} changes the goodness of fit measure in parameter estimation to mean-squared error,  for all options, see \code{\link[=cm_options]{cm_options()}}.}

\item{mem}{(optional, default: 0) A number, string, or RHS \link[stats:formula]{formula}, the prior gains or losses in memory. Formula and string refer to variables in \code{data}, for example \code{~ xoutc}.}

\item{editing}{(optional) A string, the editing rule to use (see Thaler & Johnson, 1999, pp. 645), currently only \code{"hedonic"}.}

\item{discount}{A number, how many initial trials to not use during parameter fitting.}

\item{...}{other arguments, ignored.}
}
\value{
Returns a cognitive model object, which is an object of class \href{Cm}{cm}. A model, that has been assigned to \code{m}, can be summarized with \code{summary(m)} or \code{anova(m)}. The parameter space can be viewed using \verb{pa. rspace(m)}, constraints can be viewed using \code{constraints(m)}.
}
\description{
Fits cumulative prospect theory, CPT (Tversky & Kahneman, 1992).
\itemize{
\item \code{cpt_d()} fits CPT for discrete responses = choices.
\item \code{cpt_c()} fits CPT for continuous responses = utility values.
\item \code{cpt_mem_d()} fits CPT with an editing step based on memory for discrete responses = choices (Thaler & Johnson, 1990).
\item \code{cpt_mem_c()} fits CPT with an editing step based on memory for continuous responses = utility values (Thaler & Johnson, 1990).
}
}
\details{
Fits cumulative prospect theory.
\subsection{Parameter Space}{

The model has the following free parameters:
\itemize{
\item \emph{\strong{\code{alpha}}} the utility exponent for positive outcomes.
\item \emph{\strong{\code{beta}}} the utility exponent for negative outcomes.
\item \emph{\strong{\code{gammap}}} the probability distortion for positive outcomes.
\item \emph{\strong{\code{gamman}}} the utility exponent for negative outcomes.
\item \emph{\strong{\code{lambda}}} the loss aversion.
\item In \code{cpt_d()} and \code{cpt_mem_d()}: If \code{choicerule = "softmax"}: \emph{\strong{\code{tau}}}  is the temperature or choice softness, higher values cause more equiprobable choices. If \code{choicerule = "epsilon"}: \emph{\strong{\code{eps}}} is the error proportion, higher values cause more errors from maximizing..
}
}
}
\examples{
## From Tversky, A., & Kahneman, D. (1992).
dt <- data.frame(
  x1 = c(100, -100),
  px = 1,
  x2 = 0,
  y1 = c(200, -200),
  py = c(.71,.64),
  y2 = 0,
  rp = 1)

# Make the model -------------------------------------------
# add fix parameters (don't fit)
# using the Parameter from the paper

# Discrete responses with choicerule
M <- cpt_d(rp ~ x1 + px + x2 | y1 + py + y2, ref = 0,
         choicerule = "softmax", data = dt,
         fix = list(alpha = 0.88, beta = 0.88, lambda = 2.25,
         gammap = 0.61, gamman = 0.69, tau = 1))
# View the model
M        # has a parameter `tau`

# Continuous responses/utility
M <- cpt_c(rp ~ x1 + px + x2 | y1 + py + y2, ref = 0,
         data = dt,
         fix = list(alpha = 0.88, beta = 0.88, lambda = 2.25,
         gammap = 0.61, gamman = 0.69))
# View the model
M        # No parameter `tau`

# Methods ---------------------------------------------------
predict(M, "value") # predict values, also: M$predict("value")
predict(M, "mode") # predict choice probability after softmax
summary(M)
anova(M)
}
\references{
Tversky, A., & Kahneman, D. (1992). Advances in prospect theory: cumulative representation of uncertainty. Journal of Risk and Uncertainty, 5, 297–-323. doi:\href{https://doi.org/10.1007/BF00122574}{10.1007/BF00122574}

Thaler, R. H., & Johnson, E. J. (1990). Gambling with the House Money and Trying to Break Even: The Effects of Prior Outcomes on Risky Choice. Management Science, 36(6), 643--660. doi:\href{https://doi.org/10.1287/mnsc.36.6.643}{10.1287/mnsc.36.6.643}
}
\seealso{
Other cognitive models: 
\code{\link{baseline_const_c}()},
\code{\link{bayes}()},
\code{\link{choicerules}},
\code{\link{ebm}()},
\code{\link{hm1988}()},
\code{\link{shift}()},
\code{\link{shortfall}},
\code{\link{utility}}
}
\author{
Jana B. Jarecki, \email{jj@janajarecki.com}
}
\concept{cognitive models}
