% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebm.R
\name{ebm}
\alias{ebm}
\title{Exemplar-based cognitive model}
\usage{
ebm(formula, data, criterion, fix, mode, ...)
}
\arguments{
\item{formula}{A formula (e.g., \code{class ~ f1 + f2}, or probably \code{judgment ~ dim1 + dim2 + dim3}) specifying the observed data and the feature variables.}

\item{data}{A data.frame containing \code{formula}'s variables.}

\item{criterion}{A formula (e.g., \code{~ true_class} or \code{~ true_value}) specifying the true class or criterion, note the "\code{~}".}

\item{fix}{(optional)  A list with parameter constraints (e.g., \code{list(beta = 1)}). Allowed are numbers, strings, and NA. Supplying
\code{list(alpha=0.5, beta="alpha", gamma=NA)} constrains the parameters alpha, beta, and gamma as follows
\itemize{
\item{\code{alpha=0.5}}{: fix alpha to 0.5.}
\item{\code{beta="alpha"}}{: fix beta to the value of alpha; alpha may be a free or fixed model parameter in \code{parspace}.}
\item{\code{delta=NA}}{: ignore delta if it can be ignored by setting delta equal to the value in the column "na" of \code{parspace}, given that parspace has a value in the column "na" for delta.}
}}

\item{...}{other arguments from other functions, currently ignored.}
}
\value{
An exemplar-based model; a model \code{M} can be viewed with \code{summary(M)}, or \code{anova(M)}.
}
\description{
Fits an exemplar-based model, which is also known as generalized context model (GCM, Medin & Schaffer, 1978; Nosofsky, 1986). The model predicts the value of a current stimulus from the values of previous stimuli, weighted by the similarity between the current and the previous stimuli's features. This model can be used to fit judgments (continuous) or categorization (discrete) modes.
}
\details{
If \code{type} is missing it will be inferred.
}
\section{Parameter Space of the Model}{

\itemize{
\item{n weights}{n attribute weights (range: 0-1, sum to 1), the parameter names equal the RHS of \code{formula}.}
\item{lambda}{Sensitivity (range: 0.001-10), higher values mean more discriminability within the psychological space.}
\item{r}{Exponent in the decay function (range: 1-2), a value of 1 yields exponential decay, 2 yields Gaussian decay.}
\item{q}{}
\item{m biases}{Bias towards categories (range: 0-1, sums to 1), the parameter names are \code{bx, by} were x,y are the categories}
}
}

\examples{
No examples yet
}
\references{
{Medin, D. L., & Schaffer, M. M. (1978). Context theory of classification learning. Psychological Review, 85, 207-238. doi:10.1037//0033-295X.85.3.207}

{Nosofsky, R. M. (1986). Attention, similarity, and the identification-categorization relationship. Journal of Experimental Psychology: General, 115, 39-57. doi:10.1037/0096-3445.115.1.39}

{Juslin, P., Olsson, H., & Olsson, A.-C. (2003). Exemplar effects in categorization and multiple-cue judgment. Journal of Experimental Psychology: General, 132, 133–156. doi:10.1037/0096-3445.132.1.133}
}
\author{
Jana B. Jarecki, \email{jj@janajarecki.com}
}
