% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebm.R
\name{ebm}
\alias{ebm}
\title{Exemplar-based cognitive model}
\usage{
ebm(formula, data, type, ...)
}
\arguments{
\item{formula}{a formula specifying the variables (\code{reponse ~ feature1 + feature2 + ... | criterion}), which needs a pipe (\code{|}) before the criterion variable. Examples:
\itemize{
  \item Categorization model: \code{category ~ f1 + f2 + f3 | true_class}
  \item Judgment model: \code{judgment ~ f1 + f2 + f3 | value}
  \item Memory-based preference model: \code{purchased ~ f1 + f2 + f3 | value | price} (see details)
}}

\item{data}{a data.frame or matrix containing the variables in \code{formula}; make sure it is ordered in the order in which respondents saw the stimuli.}

\item{type}{(optional) String specifying the type of model: \code{"categorization", "judgment", "valuebasedchoice"}, will be inferred if omitted.}

\item{...}{other arguments from other functions, currently ignored.}
}
\value{
An exemplar-based model; it can be viewed with \code{summary(mod)}, where \code{mod} is the name of the model object.
}
\description{
Fits an exemplar-based model (also known as generalized context model, GCM, Medin & Schaffer, 1978; Nosofsky, 1986). The model predicts the value of a current stimulus from the values of previous stimuli, weighted by the similarity between the current and the previous stimuli's features. This model can be used to fit judgments (1,2,...,100) or categorization (0,1) responses.
}
\details{
You may add a 'cost' term to the formula separated by a second pipe:  \code{reponse ~ feature1 + feature2 + ... | criterion | cost}. The model then predicts a value, compares this value to cost, and if the value is greater than the cost returns \code{1}, otherwise \code{0} (chose the stimulus, reject the stimulus). This is implemented with a softmax choice rule.
}
\examples{
#  No examples yet
}
\references{
{Medin, D. L., & Schaffer, M. M. (1978). Context theory of classification learning. Psychological Review, 85, 207-238. doi:10.1037//0033-295X.85.3.207}

{Nosofsky, R. M. (1986). Attention, similarity, and the identification-categorization relationship. Journal of Experimental Psychology: General, 115, 39-57. doi:10.1037/0096-3445.115.1.39}

{Juslin, P., Olsson, H., & Olsson, A.-C. (2003). Exemplar effects in categorization and multiple-cue judgment. Journal of Experimental Psychology: General, 132, 133–156. doi:10.1037/0096-3445.132.1.133}
}
\author{
Jana B. Jarecki, \email{jj@janajarecki.com}
}
