% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm1988.R
\name{hm1988}
\alias{hm1988}
\title{Houston & McNamara's (1988) dynamic optimization for risk-sensitive foraging problems with discrete time}
\usage{
hm1988(env, formula, data, choicerule, fixed = NULL)
}
\arguments{
\item{env}{environment}

\item{data}{data frame}
}
\description{
Houston & McNamara's (1988) dynamic optimization for risk-sensitive foraging problems with discrete time
}
