% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-hm1988.R
\name{hm1988}
\alias{hm1988}
\title{Houston & McNamara's (1988) dynamic optimization model for risk-sensitive foraging problems in discrete time}
\usage{
hm1988(
  formula,
  trials,
  states,
  budget,
  ntrials,
  initstate = 0,
  data = NULL,
  choicerule = NULL,
  fix = list(),
  options = NULL,
  fitnessfun = NULL
)
}
\arguments{
\item{formula}{A \link[stats]{formula}, e.g. \code{choice ~ x1 + px1 + x2 + px2 | y1 + py1 + y2 + py2}. Defines the risky option's variables in \code{data} by alternating outcome1 + probability1 + outcome2 + probability2 + ... Separate different options by a \code{"|"}. The Left-hand-side is optional.}

\item{trials}{(default \code{".ALL"}) Specifies the decision tiral, can be a number, numeric vector, string, formula, or \code{".ALL"}. Vectors must have length \code{nrow(data)}, formulas and strings must refer to a column in \code{data} (e.g., \code{~ time} or \code{"time"} refer to \code{data$time}). If \code{".ALL"} the model predicts for all possible trials and states.}

\item{states}{(default \code{".ALL"}) Specifies the state, can be a number, numeric vector, string, formula, or \code{".ALL"}. Vectors must have length \code{nrow(data)}, formulas must refer to a column in \code{data} (e.g., \code{~ mystate}). If \code{".ALL"} the model predicts for all possible states and trials.}

\item{budget}{A number; the goal/requirement/critical state, that matters in the terminal payout function. Can also be a numeric vector (of length \code{nrow(data)}), or a formula (e.g., \code{~b} in which case the model takes \code{data$b} as budget).}

\item{ntrials}{A number; the total number of trials available, must be greater than \code{trial}. Can also be a numeric vector of length \code{nrow(data)}, or a formula (e.g., \code{~ total}) referring to a column in \code{data} (e.g., \code{data$total}).}

\item{initstate}{(default 0) A number; the starting state in the first trial. Can also be a numeric vector (of length \code{nrow(data)}) or a formula (e.g., \code{~state0}) in which case the model takes \code{data$state0}).}

\item{data}{A data frame holding the variables of the gambles in \code{formula}.}
}
\value{
An object of class R6 holding the model, it has no free parameters. A model object \code{M} can be viewed with \code{M}, predictions can be made with \code{M$predict()} for choice predictions, and \code{M$predict("ev")} for the expected value of the optimal choice and \code{M$predict("value", 1:2)} for the expected value of all choices.
}
\description{
Houston & McNamara's (1988) dynamic optimization model for risk-sensitive foraging problems in discrete time
}
\details{
Risk-sensitive foraging means you have, for instance, four choices between the same two risky lotteries and after the four choices you need to have accumulated at least 12 points to get a reward. The optimal solution to this choice problem relies on dynamic programming. The function creates all possible future states given the possible remaining trials, and predicts the optimal choice polica or the expected value of chosing either option given a certain state and a certain time horizon.
}
\examples{
## Make fake data ---------------------------------
D <- data.frame(x1 = 0, x2 = 1, x3 = 2, px11 = 0.1, px12 = 0.8, px13 = 0.1, px21 = 0.4, px22 = 0.2, px23 = 0.4, s = rep(9:11, each = 4), init = rep(9:11, each = 4), t = 4:1)

## Setup the model --------------------------------
M <- hm1988(~ x1+px11+x2+px12+x3+px13 | x1+px21+x2+px22+x3+px23,
             trials = ~t, states = ~s, budget = 12, ntrials = 4,
             initstate = ~init, data = D, choicerule = "argmax")

M                          # View model
predict(M)                 # Predict choice probability of 1st option (arg-max)
predict(M, type="values")  # Predict expected values
}
\references{
Houston, A. I., & McNamara, J. M. (1988). A framework for the functional analysis of behaviour. Behavioural and Brain Science, 11, 117–163.
}
\author{
Jana B. Jarecki, \email{jj@janajarecki.com}
}
