% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm-methods.R
\name{npar}
\alias{npar}
\alias{npar.cm}
\title{Number of Parameters, Attributes, and Stimuli}
\usage{
npar(x)

\method{npar}{cm}(x, type = "all", ...)
}
\arguments{
\item{x}{a model object}

\item{...}{ignored}

\item{type}{A string: \code{"all"} counts all parameters, \code{"free"} counts the free parameters, \code{"fix"} countes the fixed parameters, \code{"constrained"} counts constrained parameters.}
}
\description{
\code{npar(m)} counts the parameters in a cm, \code{nstim(m)} counts stimuli, \code{natt(m)} counts attributes by stimulus, \code{nobs(m)} counts observations in the data of a model.
}
\examples{
D <- data.frame(x = 1, y = 1, z = 1)
M <- bayes_beta(y ~ x + z, D, fix = "start")
npar(M) # 3
npar(M)
}
