% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsenvironment-class.R
\docType{data}
\name{Rsenvironment}
\alias{Rsenvironment}
\title{Class for risk-sensitive foraging environments}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Rsenvironment
}
\description{
Class for risk-sensitive foraging environments
}
\keyword{datasets}
