% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsfft.R
\name{rsfft}
\alias{rsfft}
\title{Risk-sensitive foraging fast and frugal tree}
\usage{
rsfft(formula = NULL, sbt = NULL, nopt = NULL, nout = NULL,
  fixed = NULL, data = NULL, env = NULL, choicerule,
  terminal.fitness.fun)
}
\arguments{
\item{formula}{Formula (e.g., \code{y ~ x1 + x2}); depends on the model.}

\item{fixed}{(optional) List of model parameters and their fixed values, those parameters won't be estimated. Fixing a parameter to \code{NA} means ignoring it entirely (see details).}

\item{data}{Data.frame or matrix holding \code{formula}'s variables.}

\item{choicerule}{(optional, default \code{NULL}) String specifying the choice rule. Allowed values: \code{NULL} or e.g., \code{"softmax"}, values of \code{type}s in \link[cogsciutils]{choicerule}.}
}
\description{
Risk-sensitive foraging fast and frugal tree
}
\keyword{internal}
