% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsfft.R
\name{rsfft}
\alias{rsfft}
\title{Risk-sensitive foraging fast and frugal tree}
\usage{
rsfft(
  formula = NULL,
  sbt = NULL,
  nopt = NULL,
  nout = NULL,
  fix = NULL,
  data = NULL,
  env = NULL,
  choicerule,
  terminal.fitness.fun
)
}
\description{
rsfft()
}
\keyword{internal}
