% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shortfall_cpp}
\alias{shortfall_cpp}
\title{Computes Predictions in the Shortfall Model}
\usage{
shortfall_cpp(x, p, a, beta, delta)
}
\arguments{
\item{x}{Numeric matrix, the outcomes}

\item{p}{Numeric matrix, the probabilities}

\item{a}{Numeric vector, the aspiration levels}

\item{beta}{Model parameter}

\item{delta}{Model parameter}
}
\description{
Computes Predictions in the Shortfall Model
}
