% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-treshold.R
\name{threshold}
\alias{threshold}
\alias{threshold_c}
\alias{threshold_d}
\title{Threshold Model}
\usage{
threshold(
  formula,
  data,
  fix = list(),
  choicerule = NULL,
  mode,
  discount = 0L,
  options = list(),
  ...
)

threshold_c(
  formula,
  data,
  fix = list(),
  choicerule = NULL,
  discount = 0,
  options = list(),
  ...
)

threshold_d(
  formula,
  data,
  fix = list(),
  choicerule = "softmax",
  discount = 0,
  options = list(),
  ...
)
}
\arguments{
\item{formula}{A formula specifying choice ~ var}

\item{...}{other arguments from other functions, currently ignored.}
}
\value{
A model object (similar to lm-objects) of class "treshold". It can be viewed with \code{summary(mod)}, where \code{mod} is the name of the model object.
}
\description{
\code{treshold()} fits a threshold model with the threshold as free parameter, \code{treshold_c()} predicts the distance to a threshold, \code{treshold_d()} predicts a discrete response given the distance to a threshold.
}
\details{
Given the formula \code{y ~ a} the model predicts y = 1 for a >= \code{nu} and y = 0 for a < \code{nu}
}
\section{Parameter Space}{

\tabular{lrcllr}{\verb{   }\strong{Name} \tab \verb{    }\strong{LB} \tab  \strong{-} \tab \strong{UB}\verb{    } \tab \strong{Description} \tab \strong{Start Value}\cr
\verb{   }\code{nu} \tab  -Inf \tab  - \tab  Inf \tab  Treshold \tab  0
}
}

\examples{
D <- data.frame(
       y = rep(0:1, each=5),
       a = 1:10)

M <- threshold_c(y ~ a, D, fix="start")        # fixed par. to start values
predict(M)                                     # predict dist. to threshold
anova(M)                                       # anova-like table
summary(M)                                     # summarize

M <- threshold_d(y ~ a, D, fix="start")        # fixed par. to start values
predict(M)                                     # predict dist. to threshold
anova(M)                                       # anova-like table
summary(M)                                     # summarize
M$MSE()                                        # mean-squared error   

### Binary response given a threshold
# --------------------------------------------
M <- threshold(y ~ a, D, fix="start", choicerule = "softmax")
predict(M)                       #  --"--  maximum posterior
anova(M)                                       # anova-like table
summary(M)                                     # summarize
M$MSE()                                        # mean-squared error    


### Parameter specification and fitting
----------------------------------------
# Use a response variable, y, to which we fit parameter
threshold(y ~ a, D, fix = "start", "softmax")     # "start" fixes all par.,
                                                  # and fits none 
threshold(y ~ a , D, list(nu=2), "softmax")       # fix threshold nu to 2
threshold(y ~ a, D, list(tau=0.5), "softmax")     # fix soft-max tau to 1   
threshold(y ~ a, D, choicerule = "softmax")       # nu and tau free param   


}
\author{
Jana B. Jarecki
}
