% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-utility.R
\name{utility_pow}
\alias{utility_pow}
\alias{utility}
\title{Utility Function Models}
\usage{
utility_pow(
  formula,
  data,
  fix = list(),
  choicerule = NULL,
  mode = NULL,
  discount = 0,
  options = list(),
  ...
)

utility(
  formula,
  data,
  type = c("power"),
  fix = list(),
  choicerule = NULL,
  mode = NULL,
  discount = 0,
  options = list(),
  ...
)
}
\arguments{
\item{formula}{A formula specifying responses ~ values (e.g, \code{y ~ x1 | x2}).}

\item{type}{A string. Currently only \code{"power"} for power utility.}
}
\value{
A model object (similar to lm-objects) of class "utility", which can be viewed with \code{summary(M)} or \code{anova(M)} and predictions can be made with \code{predict(M)}.
}
\description{
\code{utility()} fits utility models, e.g., \code{utility_pow()} fits a power utility
}
\details{
\emph{Power Utility [(\code{utility(type="power"), utility_pow()}].} The utility \eqn{U(x)} for positive outcomes \eqn{x} is \eqn{x^r if r > 0}, and is \eqn{log(x) if r = 0}, and is \eqn{-x^r if r < 0}. The utility for negative outcomes \eqn{x} equals \eqn{-U(-x)} with a separate exponent \code{rn}.
}
\section{Parameter Space}{

\tabular{lcrcllr}{\verb{   } \tab \strong{Name} \tab \verb{    }\strong{LB} \tab  \strong{-} \tab \strong{UB}\verb{    } \tab \strong{Description} \tab \strong{Start Value}\cr

\verb{   } Power \tab \code{rp} \tab  \eqn{-}20* \tab  - \tab  20 \tab  Exponent for positive \eqn{x \ge 0}. Values < 0 yield concavity, 0 logarithm, > 0 convexity \tab  1\cr

\verb{   } \tab \code{rn} \tab  \eqn{-}20* \tab  - \tab 20 \tab  Exponent for negative \eqn{x < 0;}\tab  1
}
\verb{   }*\emph{Note}, the lower bound is 0.001 if \eqn{x} contains positive \emph{and} negative values (see Wakker, 2008).
}

\examples{
#  No examples yet

}
\references{
{Wakker, P. P. (2008). Explaining the characteristics of the power (CRRA) utility family. \emph{Health Economics, 17(12), 1329-1344. } \url{htrps://doi.org/10.1002/hec.1331}}

{Tversky, A. (1967). Utility theory and additivity analysis of risky choices. \emph{Journal of Experimental Psychology, 75(1)}, 27-36. \url{htrp://dx.doi.org/10.1037/h0024915}}
}
