% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-utility.R
\name{utility}
\alias{utility}
\alias{utility_pow_d}
\alias{utility_pow_c}
\title{Utility Function Models}
\usage{
utility_pow_d(
  formula,
  data,
  choicerule,
  fix = list(),
  discount = 0,
  options = list(),
  ...
)

utility_pow_c(formula, data, fix = list(), discount = 0, options = list(), ...)
}
\arguments{
\item{formula}{A \link[stats:formula]{formula}, the variables in \code{data} to be modeled. For example, \code{y ~ x1} models response y as function of one stimulus value x1.}

\item{data}{A data frame, the data to be modeled.}

\item{choicerule}{A string, the choice rule. Allowed values, see  \code{cm_choicerules()}: \code{"none"} is no choice rule, \code{"softmax"} is \href{https://en.wikipedia.org/wiki/Softmax_function}{soft-maximum}, \code{"luce"} is \href{https://en.wikipedia.org/wiki/Luce\%27s_choice_axiom}{Luce's axiom}.}

\item{fix}{(optional) A list with parameter-value pairs of fixed parameters. If missing all free parameters are estimated. If set to \code{"start"} all parameters are fixed to their start values. Model parameter names are \emph{\code{rp}, \code{rn}} (see details - model parameters).
\itemize{
\item \code{list(rp = 5.40)} sets parameter \emph{\code{rp}} equal to 5.40.
\item \code{list(rp = "rn")} sets parameter \emph{\code{rp}} equal to parameter \emph{\code{rn}} (estimates \emph{\code{rn}}).
\item \code{list(rn = "rp", rp = 5.40)} sets parameter \emph{\code{rn}} equal to parameter \emph{\code{rp}} and sets \emph{\code{rp}} equal to 5.40 (estimates none of the two).
\item \code{list(rp = NA)} omits the parameter \emph{\code{rp}}, if possible.
\item \code{"start"} sets all parameters equal to their initial values (estimates none). Useful for building a first test model.
}}

\item{discount}{A number, how many initial trials to not use during parameter fitting.}

\item{options}{(optional) A list, list entries change the modeling procedure. For example, \code{list(lb = c(k=0))} changes the lower bound of parameter \emph{k} to 0, or \code{list(fit_measure = "mse")} changes the goodness of fit measure in parameter estimation to mean-squared error,  for all options, see \link{cm_options}.}

\item{...}{other arguments, ignored.}
}
\value{
Returns a cognitive model object, which is an object of class \href{Cm}{cm}. A model, that has been assigned to \code{m}, can be summarized with \code{summary(m)} or \code{anova(m)}. The parameter space can be viewed using \verb{pa. rspace(m)}, constraints can be viewed using \code{constraints(m)}.
}
\description{
Fits utility models.
\itemize{
\item \code{utility_pow_c()} fits a power utility for continuous responses.
\item \code{utility_pow_d()} fits a power utility for discrete respoonses.
}
}
\details{
The power utility \eqn{U(x)} of positive inputs, \eqn{x > 0}, is \eqn{x^r} if \eqn{r > 0}, and is \eqn{log(x)} if \eqn{r = 0}, and is \eqn{-x^r} if \eqn{r < 0}. The power utility of negative inputs \eqn{x} is \eqn{-U(-x)} with a separate exponent r (Wakker, 2008). To fit a power utility with one single exponent for positive and negative \emph{x}, set \code{fix = list(rp = "rn")}, not recommended for mixed input.
\subsection{Model Parameters}{

The model has between 1 and 3 free parameters, depending on model and \code{data} (see \code{\link[=npar]{npar()}}):
\itemize{
\item \emph{\strong{\code{rp}}} is the power utility exponent for positive data \eqn{x \ge} 0 (omitted if all \eqn{x <} 0).
\item \emph{\strong{\code{rn}}} is the exponent for negative data \eqn{x < 0} (omitted if all \eqn{x \ge} 0).
\item In \code{utility_pow_c()}: \emph{\strong{\code{sigma}}} is the standard deviation of the normally-distributed loglikelihood of the responses.
\item In \code{utility_pow_d()}: If \code{choicerule = "softmax"}: \emph{\strong{\code{tau}}}  is the temperature or choice softness, higher values cause more equiprobable choices. If \code{choicerule = "epsilon"}: \emph{\strong{\code{eps}}} is the error proportion, higher values cause more errors from maximizing.
}
}
}
\examples{
#  No examples yet
}
\references{
{Wakker, P. P. (2008). Explaining the characteristics of the power (CRRA) utility family. \emph{Health Economics, 17(12)}, 1329-1344. doi:\href{htrps://doi.org/10.1002/hec.1331}{10.1002/hec.1331}}

{Tversky, A. (1967). Utility theory and additivity analysis of risky choices. \emph{Journal of Experimental Psychology, 75(1)}, 27-36. doi:\href{htrp://dx.doi.org/10.1037/h0024915}{10.1037/h0024915}}
}
\seealso{
Other cognitive models: 
\code{\link{baseline_const_c}()},
\code{\link{bayes}()},
\code{\link{choicerules}},
\code{\link{cpt}},
\code{\link{ebm}()},
\code{\link{hm1988}()},
\code{\link{shift}()},
\code{\link{shortfall}},
\code{\link{threshold}()}
}
\author{
Jana B. Jarecki, \email{jj@janajarecki.com}
}
\concept{cognitive models}
