% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-utility.R
\name{utility}
\alias{utility}
\alias{utility_pow_d}
\alias{utility_pow_c}
\title{Utility Function Models}
\usage{
utility_pow_d(
  formula,
  data,
  choicerule,
  fix = list(),
  discount = 0,
  options = list(),
  ...
)

utility_pow_c(formula, data, fix = list(), discount = 0, options = list(), ...)
}
\arguments{
\item{formula}{A formula, the variables in data to be modeled. For example, \code{y ~ x | z} models the response \code{y} as function of a stimulus value \code{x} and \code{z}. Lines (|) separate stimuli.}

\item{data}{A data frame.}

\item{choicerule}{(discrete responses only) A string, the choice rule.  \code{cm_choicerules()} shows the allowed values. \code{"none"} is no choice rule, \code{"softmax"} is \href{https://en.wikipedia.org/wiki/Softmax_function}{soft-maximum}, \code{"luce"} is \href{https://en.wikipedia.org/wiki/Luce\%27s_choice_axiom}{Luce's axiom}.}

\item{fix}{(optional) A list or the string \code{"start"}, the model parameters that are fixed and not estimated. Model parameters are called \emph{\code{rp}, \code{rn}} (see details), if this argument is empty all parameters are estimated.
\itemize{
\item \code{"start"} sets all parameters equal to their initial values (estimates none). Useful for building a first test model.
\item \code{list(rp = 5.40)} sets parameter \emph{\code{rp}} equal to 5.40.
\item \code{list(rp = "rn")} sets parameter \emph{\code{rp}} equal to parameter \emph{\code{rn}} (estimates \emph{\code{rn}}).
\item \code{list(rn = "rp", rp = 5.40)} sets parameter \emph{\code{rp}} equal to parameter \emph{\code{rn}} and sets parameter \emph{\code{rn}} equal to 5.40 (estimates none of the two).
}}

\item{options}{(optional) A list to change the parameter estimation process, see \code{\link[=cm_options]{cm_options()}} or the section Options below.}

\item{type}{(optional) A string, utility function, currently only \code{"power"} for power utility.}
}
\value{
Returns a cognitive model object, which is an object of class \href{Cm}{cm}. A model, that has been assigned to \code{m}, can be summarized with \code{summary(m)} or \code{anova(m)}. The parameter space can be viewed using \code{parspace(m)}, constraints can be viewed using \code{constraints(m)}.
}
\description{
\code{utility()} fits utility models.
\itemize{
\item \code{utility_pow_c()} fits a power utility for continuous responses.
\item \code{utility_pow_d()} fits a power utility for discrete respoonses.
}
}
\details{
The power utility \eqn{U(x)} of positive inputs, \eqn{x > 0}, is \eqn{x^r} if \eqn{r > 0}, and is \eqn{log(x)} if \eqn{r = 0}, and is \eqn{-x^r} if \eqn{r < 0}. The power utility of negative inputs \eqn{x} is \eqn{-U(-x)} with a separate exponent r (Wakker, 2008). To fit a power utility with one single exponent for positive and negative \emph{x}, set \code{fix = list(rp = "rn")}, not recommended for mixed input.
\subsection{Model Parameters}{

The model has between 1 and 3 free parameters, depending on model and \code{data} (see \code{\link[=npar]{npar()}}):
\itemize{
\item \emph{\strong{\code{rp}}} is the power utility exponent for positive data \eqn{x \ge} 0 (omitted if all \eqn{x <} 0).
\item \emph{\strong{\code{rn}}} is the exponent for negative data \eqn{x < 0} (omitted if all \eqn{x \ge} 0).
\item In \code{utility_pow_c()}: \emph{\strong{\code{sigma}}} is the standard deviation of the normally-distributed loglikelihood of the responses.
\item In \code{utility_pow_d()}:  If \code{choicerule = "softmax"}: \emph{\strong{\code{tau}}}  is the temperature or choice softness, higher values cause more equiprobable choices. If \code{choicerule = "epsilon"}: \emph{\strong{\code{eps}}} is the error proportion, higher values cause more errors from maximizing.
}
}

\subsection{Options}{

\verb{   } (Optional) Set \code{options} to fine-tune the parameter estimation
process. The values listed below must be supplied in list. For example
\code{options = list(lb = c(k = -10))} changes a lower parameter bound.

\describe{
\item{\verb{   }\code{lb}}{Named numeric vector, changes lower parameter bounds;
\code{lb = c(k = -10)} lets parameter \emph{k} start at -10.}
\item{\verb{   }\code{ub}}{Named numeric vector, changes upper parameter bounds:
\code{ub = c(k = 10)} lets parameter \emph{k} go until 10.}
\item{\verb{   }\code{start}}{Named numeric vector, changes parameter start
values: \code{start = c(k = 5)} lets parameter \emph{k} start at 5.}
\item{\verb{   }\code{fit}}{Logical (default \code{TRUE}), \code{fit = FALSE} disables
parameter estimation. Useful for testing models.}
\item{\verb{   }\code{fit_measure}}{A string (default \code{"loglikelihood"}), the
\href{https://en.wikipedia.org/wiki/Goodness_of_fit}{goodness of fit}
measure to be optimized during parameter estimation.
Can be any value of \code{type} in \code{\link[cognitiveutils:gof]{cognitiveutils::gof()}}.
Loglikelihood uses a binomial PDF in models with discrete
data and a normal PDF for continuous data, \eqn{N(\mu, \sigma)}
with  \eqn{\mu}=predictions, \eqn{\sigma}=constant,
estimated as additional free paramter. Change the PDF using
\code{fit_args = list(pdf = "...")}}
\item{\verb{   }\code{fit_args}}{Named list, options for parameter estimation.
Can be any
argument to \code{\link[cognitiveutils:gof]{cognitiveutils::gof()}}. For example,
\code{list(pdf = "truncnorm", a = 0, b = 1)} changes the PDF
in the log likelihood to a truncated normal between 0 and 1,
\code{list(pdf = "multinom")} changes it to a multinomial PDF.
\code{list(grid_offset = .01)} offsets the parameter in a grid search
by 0.01 from the parameter boundaries, \code{list(nsteps = 10)} defines
10 steps for each parameter in the regular grid in the grid search.}
\item{\verb{   }\code{fit_data}}{A data frame, the data to estimate the model parameters
from. Needed if the data for the parameter estimation differs from
the data in the main \code{data} argument in a model.}
\item{\verb{   }\code{solver}}{A string, the optimizer for the parameter estimation. Run
\code{cm_solvers()} to list all solvers. Can be \code{"grid"}, \code{"solnp"}
\code{"optimx"}, \code{"nloptr"}, \code{"nlminb"} and others from \link{ROI}. Can be
\code{c("grid", "xxx")}: a grid-plus-optimization: A grid search, followed
by an optimization with xxx using the \emph{n}
best solutions as start values in the optimization; the overal best
parameter set wins; and \emph{n} can be set in \code{solver_args$nbest}
Changing the solver may cause warnings and ignored parameter bounds.}
\item{\verb{   }\code{solver_args}}{A named list, additional arguments passed directly
to the solver function, see the pages of the solver to see which
arguments the solver function has. For example:
\code{list(offset = 0.01)} offset the parameters from their boundaries
when \code{solver = "grid"}.
\code{list(nsteps = 10)} uses 10 steps for each parameter in the regular
grid, for \code{solver = "grid"}), \code{list(nbest = 3)} uses the 3 best
parameter sets from the grid search as starting values in a
grid-plus-optimization solver, for \code{solver = c("grid", "xxx")}.
\code{list(control = )} control arguments in the solver
\href{Rsolnp::solnp()}{solnp} and the
\href{https://rdrr.io/cran/ROI/man/ROI_solve.html}{ROI solvers}}
}
}
}
\examples{
#  No examples yet
}
\references{
{Wakker, P. P. (2008). Explaining the characteristics of the power (CRRA) utility family. \emph{Health Economics, 17(12)}, 1329-1344. doi:\href{htrps://doi.org/10.1002/hec.1331}{10.1002/hec.1331}}

{Tversky, A. (1967). Utility theory and additivity analysis of risky choices. \emph{Journal of Experimental Psychology, 75(1)}, 27-36. doi:\href{htrp://dx.doi.org/10.1037/h0024915}{10.1037/h0024915}}
}
\seealso{
Other cognitive models: 
\code{\link{Cm}},
\code{\link{baseline_const_c}()},
\code{\link{bayes}()},
\code{\link{gcm}()}
}
\concept{cognitive models}
