% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-utility.R
\name{utility_pow_d}
\alias{utility_pow_d}
\alias{utility_pow_c}
\title{Utility Function Models}
\usage{
utility_pow_d(
  formula,
  data,
  fix = list(),
  choicerule,
  discount = 0,
  options = list(),
  ...
)

utility_pow_c(formula, data, fix = list(), discount = 0, options = list(), ...)
}
\arguments{
\item{formula}{A formula specifying responses ~ values (e.g, \code{y ~ x1 | x2}).}

\item{type}{(optional) A string, which utility function to use; currently available is only \code{"power"} for power utility.}
}
\value{
A model object of class "utility", which can be viewed with \code{summary(.)} or \code{anova(.)}; predictions can be made with \code{predict(.)}.
}
\description{
\code{utility()} fits utility models.
\itemize{
\item \code{utility_pow_c()} fits a power utility for continuous responses.
\item \code{utility_pow_d()} fits a power utility for discrete respoonses.
}
}
\details{
\emph{Power Utility} \strong{\code{utility_pow_.()}}. The utility \eqn{U(x)} for positive outcomes, \eqn{x > 0}, is \eqn{x^r if r > 0}, and is \eqn{log(x) if r = 0}, and is \eqn{-x^r if r < 0}. The utility for negative outcomes \eqn{x} equals \eqn{-U(-x)} with a separate exponent r (Wakker, 2008). The exponent is called \code{rp} and \code{rn} for positive and negative outcomes, respectively. To fit the model with only one exponent parameter, which is not recommended for mixed outcomes, set \code{fix = list(rp = "rn")}.
}
\section{Parameter Space}{

\tabular{lcrcllr}{\verb{   } \tab \strong{Name} \tab \verb{    }\strong{LB} \tab  \strong{-} \tab \strong{UB}\verb{    } \tab \strong{Description} \tab \strong{Start Value}\cr

\verb{   } Power \tab \code{rp} \tab  \eqn{-}20* \tab  - \tab  20 \tab  Exponent for positive \eqn{x \ge 0}. Values < 0 yield concavity, 0 logarithm, > 0 convexity \tab  1\cr

\verb{   } \tab \code{rn} \tab  \eqn{-}20* \tab  - \tab 20 \tab  Exponent for negative \eqn{x < 0;}\tab  1
}
\verb{   }*\emph{Note}, the lower bound is 0.001 if \eqn{x} contains positive \emph{and} negative values (see Wakker, 2008).
}

\examples{
#  No examples yet

}
\references{
{Wakker, P. P. (2008). Explaining the characteristics of the power (CRRA) utility family. \emph{Health Economics, 17(12)}, 1329-1344. doi:\href{htrps://doi.org/10.1002/hec.1331}{10.1002/hec.1331}}

{Tversky, A. (1967). Utility theory and additivity analysis of risky choices. \emph{Journal of Experimental Psychology, 75(1)}, 27-36. doi:\href{htrp://dx.doi.org/10.1037/h0024915}{10.1037/h0024915}}
}
